% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidomics.R
\name{add_lipid_class_transition}
\alias{add_lipid_class_transition}
\title{Retrieve lipid name, lipid class and transition from feauture names}
\usage{
add_lipid_class_transition(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
MidarExperiment object
}
\description{
Retrieve lipid name, lipid class and transition from feauture names
}
