% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-read.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata an MidarExperiment object}
\usage{
add_metadata(data, metadata, excl_unannotated_analyses = FALSE)
}
\arguments{
\item{data}{MidarExperiment object}

\item{excl_unannotated_analyses}{Exclude analyses (samples) that have no matching metadata}

\item{path}{List of tibbles or data.frames containing analysis, feature, istd, response curve tables}
}
\value{
metadata list
}
\description{
Metadata provided as a list of tibbles will validates for consistency again loaded analysis data of the provided MidarExperiment object and then transfered.
}
