% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-managment.R
\name{analyses_exclude}
\alias{analyses_exclude}
\title{@title Exclude analyses from the dataset}
\usage{
analyses_exclude(data, analyses_exlude, overwrite)
}
\arguments{
\item{data}{MidarExperiment object}

\item{analyses_exlude}{Vector of analysis IDs to exclude from the dataset}

\item{overwrite}{If \code{TRUE} then existing valid_analysis flags will be overwritten, otherwise appended}
}
\value{
\code{MidarExperiment} object
}
\description{
@title Exclude analyses from the dataset
}
