% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{analysis_type,MidarExperiment-method}
\alias{analysis_type,MidarExperiment-method}
\title{Get \code{analysis_type}}
\usage{
\S4method{analysis_type}{MidarExperiment}(x)
}
\arguments{
\item{x}{MidarExperiment object}
}
\value{
A character string
}
\description{
Get \code{analysis_type}
}
