% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{analysis_type<-,MidarExperiment-method}
\alias{analysis_type<-,MidarExperiment-method}
\title{Set \code{analysis_type}}
\usage{
\S4method{analysis_type}{MidarExperiment}(x) <- value
}
\arguments{
\item{x}{MidarExperiment object}

\item{value}{Analysis type, one of "lipidomics", "metabolomics, "quantitative"}
}
\description{
Set \code{analysis_type}
}
