% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{analysis_type<-}
\alias{analysis_type<-}
\title{Get analysis type}
\usage{
analysis_type(x) <- value
}
\arguments{
\item{x}{MidarExperiment object}

\item{value}{Analysis type, one of "lipidomics", "metabolomics, "quantitative"}
}
\description{
Get the analysis type defined for the MidarExperiment object
}
