% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{analysis_type}
\alias{analysis_type}
\title{Set analysis type}
\usage{
analysis_type(x)
}
\arguments{
\item{x}{MidarExperiment object}
}
\value{
A character string
}
\description{
Set the analysis type, i.e. "lipidomics", "metabolomics, "quantitative"
}
