% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{apply_qc_filter}
\alias{apply_qc_filter}
\title{QC-filter dataset}
\usage{
apply_qc_filter(
  data,
  Intensity_BQC_min = NA,
  CV_BQC_max = NA,
  Intensity_TQC_min = NA,
  CV_TQC_max = NA,
  SB_RATIO_min = NA,
  R2_min = NA,
  RQC_CURVE = NA
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{Intensity_BQC_min}{Intensity_BQC_min}

\item{CV_BQC_max}{= CV_BQC_max}

\item{Intensity_TQC_min}{= Intensity_TQC_min}

\item{CV_TQC_max}{= CV_TQC_max}

\item{SB_RATIO_min}{= SB_RATIO_min}

\item{R2_min}{= R2_min}

\item{RQC_CURVE}{= RQC_CURVE}
}
\description{
QC-filter dataset
}
