% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midar-batchcorr.R
\name{batch_corr_center}
\alias{batch_corr_center}
\title{Combines a list of MidarExperiments into one}
\usage{
batch_corr_center(data, qc_types, center_by_fun = "median")
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types used for batch correction}

\item{center_by_fun}{Function used to center. Default is "median".}
}
\description{
Combines a list of MidarExperiments into one
}
