% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-istd-normalization.R
\name{calc_quant_by_istd}
\alias{calc_quant_by_istd}
\title{Calculate analyte concentrations}
\usage{
calc_quant_by_istd(
  data = NULL,
  error_missing_info = TRUE,
  ignore_unused_istds = TRUE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{error_missing_info}{Raise error when 1 or more ISTD concentratios and sample/ISTD amounts are missing. Default: \code{TRUE}.}

\item{ignore_unused_istds}{Ignore ISTDs with missing concentrations that are not used in any feature quantitation. Default: \code{FALSE}.}
}
\value{
MidarExperiment object
}
\description{
Calculation is based on ISTD-normalized intensities and corresponding
sample and spiked-in ISTD amounts. The determined concentration unit is based
on the \code{sample_amount_unit} provided in the analysis metadata.
If the spiked-in ISTD concentrations are missing, the concentrations of the
}
