% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{calculate_qc_metrics}
\alias{calculate_qc_metrics}
\title{Calculate QC metrics for each feature}
\usage{
calculate_qc_metrics(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
MidarExperiment object
}
\description{
Calculate QC metrics for each feature
}
