% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{$,MidarExperiment-method}
\alias{$,MidarExperiment-method}
\title{Getter for specific slots of an MidarExperiments object}
\usage{
\S4method{$}{MidarExperiment}(x, name)
}
\arguments{
\item{x}{MidarExperiment object}

\item{name}{MidarExperiment slot}
}
\value{
Value with a variable or a tibble
}
\description{
$ syntax can be used to as a shortcut for getting specific variables and results from a MidarExperiment object
}
\examples{
mexp <- MidarExperiment()
mexp$analysis_type
mexp$annot_analyses
}
