% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{check_data}
\alias{check_data}
\title{Check integrity of MidarExperiment data object}
\usage{
check_data(data = NULL)
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
silent on success, prints abort message on fail
}
\description{
Helper function that checks the structure and contents of
a MidarExperiment object
}
