% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-outlier.R
\name{clear_outlier}
\alias{clear_outlier}
\title{Clear all analysis/sample outlier classifications}
\usage{
clear_outlier(data)
}
\arguments{
\item{data}{description}
}
\value{
MidarExperiment object
}
\description{
MidarExperiment object
}
