% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_drift-batch.R
\name{corr_batcheffects}
\alias{corr_batcheffects}
\title{Batch centering}
\usage{
corr_batcheffects(
  data,
  qc_types,
  correct_location = TRUE,
  correct_scale = FALSE,
  ...
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types used for batch correction}

\item{use_raw_concs}{Apply to unadjusted (raw) feature_conc. Default is FALSE, which means previously drift-corrected concs will be used if available, otherwise unadjusted concs will be used}

\item{center_fun}{Function used to center. Default is "median".}
}
\value{
MidarExperiment object
}
\description{
A short description...
#' @details
Additional details...
}
