% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorr.R
\name{corr_drift_fun}
\alias{corr_drift_fun}
\title{Drift Correction by Custom Function}
\usage{
corr_drift_fun(
  data,
  smooth_fun,
  qc_types,
  log2_transform = TRUE,
  span_width,
  within_batch,
  apply_conditionally,
  apply_conditionally_per_batch = TRUE,
  min_sample_cv_ratio_before_after = 1,
  feature_list = NULL
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{smooth_fun}{Function that performs drift correction. Function need to have following parameter \code{data} (\code{MidarExperiment}), \code{QC_TYPES} (one or more strings), and \code{span_width} (numerical).
Function needs to return a numerical vector with the length of number of rows in \code{data}. In case functions fails a vector with NA_real_ needs be returned}

\item{qc_types}{QC types used for drift correction}

\item{log2_transform}{log2 transform data during correction (Default is TRUE). Log transformation is only used for the fit, results will not be transformed.}

\item{span_width}{Width used by smoothing function}

\item{within_batch}{Correct each batch separately (Default is TRUE)}

\item{apply_conditionally}{Apply drift correction to all species or conditionally based on 'min_sample_cv_ratio_before_after'}

\item{apply_conditionally_per_batch}{Apply correction conditionally using min_sample_cv_ratio_before_after criteriaper batch or across batches}

\item{min_sample_cv_ratio_before_after}{Maximum sample CV change for correction to be applied}

\item{feature_list}{Apply correction only to species matching (RegEx).}
}
\value{
MidarExperiment object
}
\description{
Function to correct for run-order drifts within or across batches via a provided custom function
#' @details
The drift correction function needs to be provided by the user. See \code{smooth_fun} for details.
}
