% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorr.R
\name{corr_drift_loess}
\alias{corr_drift_loess}
\title{Drift Correction by LOESS Smoothing}
\usage{
corr_drift_loess(
  data,
  qc_types,
  within_batch,
  apply_conditionally,
  apply_conditionally_per_batch = TRUE,
  log2_transform = TRUE,
  span = 0.75,
  feature_list = NULL,
  min_sample_cv_ratio_before_after = 1
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types used for drift correction}

\item{within_batch}{Correct each batch separately (Default is TRUE)}

\item{apply_conditionally}{Apply drift correction to all species or conditionally based on 'min_sample_cv_ratio_before_after'}

\item{apply_conditionally_per_batch}{Apply correction conditionally using min_sample_cv_ratio_before_after criteriaper batch or across batches}

\item{log2_transform}{log2 transform data during correction (Default is TRUE). Log transformation is only used for the fit, results will not be transformed.}

\item{span}{Loess span width (default is 0.75)}

\item{feature_list}{Apply correction only to species matching (RegEx).}

\item{min_sample_cv_ratio_before_after}{Maximum sample CV change for correction to be applied}
}
\value{
MidarExperiment object
}
\description{
Function to correct for run-order drifts within or across batches using loess smoothing
#' @details
}
