% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-isotope.R
\name{correct_interference_manually}
\alias{correct_interference_manually}
\title{Manual correction for interference contributed by another feature}
\usage{
correct_interference_manually(
  data,
  variable,
  feature,
  interfering_feature,
  relative_contribution,
  new_feature_id = NULL
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable}{Default: \code{feature_intensity}. Name of Variable to be corrected.}

\item{feature}{Name of feature to be corrected}

\item{interfering_feature}{Name of feature that is interfering, i.e. contributing to the signal of \code{feature}}

\item{relative_contribution}{Relative portion of the interfering feature to contribute to the feature signal. Must be between 0 and 1.}

\item{new_feature_id}{Optional. New name of corrected feature. If empty then feature name will not change.}
}
\value{
MidarExperiment object
}
\description{
The interference is substracted using following formula:
\deqn{Value_{Corrected} = Value_{Feature} - Factor_{Contribution} * Value_{Interfering Feature}}
}
