% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefvariation.R
\name{cvlog}
\alias{cvlog}
\title{Coefficient of variation (CV) using log-transformed data}
\usage{
cvlog(x, na.rm)
}
\arguments{
\item{x}{a numeric vector with values that are NOT log-transformed}

\item{na.rm}{logical, if TRUE then NA values are stripped from x before computation takes place}
}
\value{
numeric vector of length one. If x contains a zero, then NaN is returned. If x is not numeric or integer, NA_real_ is returned
}
\description{
Computes the coefficient of variation (CV) after log-transformation of the provided data
}
\examples{
library(midar)
cvlog(c(5,6,3,4,5, NA), na.rm = TRUE)

}
