% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-managment.R
\name{data_exclude_features}
\alias{data_exclude_features}
\title{Exclude features from the dataset}
\usage{
data_exclude_features(data = NULL, features_exlude, replace_existing)
}
\arguments{
\item{data}{MidarExperiment object}

\item{features_exlude}{Vector of feature IDs (case-sensitive) to exclude from the dataset}

\item{replace_existing}{If \code{TRUE} then existing valid_feature flags will be overwritten, otherwise appended}
}
\value{
\code{MidarExperiment} object
}
\description{
Exclude features from the dataset
}
