% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{import_data_masshunter}
\alias{import_data_masshunter}
\title{Imports Agilent MassHunter Quantitative Analysis CSV files}
\usage{
import_data_masshunter(
  data = NULL,
  path,
  include_metadata,
  file_format = "csv",
  expand_qualifier_names = TRUE,
  silent = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{One or more file paths, or a directory path (in which case all matching files will be imported)}

\item{include_metadata}{Logical, whether to extract and add metadata from the analysis result file}

\item{expand_qualifier_names}{Logical, whether to add the quantifier name in front of the qualifier name (the latter only has the m/z transition values)}

\item{silent}{Logical, whether to suppress most notifications}
}
\value{
MidarExperiment object with the imported data
}
\description{
Imports .csv files exported from Agilent MassHunter Quantitative Analysis software, containing peak integration results.
The input files must have anlyses (samples) in rows, features/compounds in columns, and either peak areas, peak heights, or response as the values.
Additional columns, such as retention time (RT), full-width at half-maximum (FWHM), precursor m/z (PrecursorMZ), and collision energy (CE), will also be imported and made available in the \code{MidarExperiment} object for downstream analyses.

When a directory path is provided, all matching .csv files in that directory will be imported and merged into a single dataset. This is useful when importing datasets that were pre-processed in blocks, resulting in multiple files.
Each unique combination of feature and raw data file must only occur once across all source data files. Duplicate combinations will result in an error.
}
\examples{
mexp <- MidarExperiment()
mexp <- import_data_masshunter(data = mexp, input_path = system.file("extdata", "Example_MHQuant_1.csv", package = "midar"), include_metadata = TRUE)
mexp <- import_data_masshunter(data = mexp, input_path = "/path/to/directory/with/masshunter/files", include_metadata = TRUE)
}
