% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-runorder.R
\name{data_plot_runsequence}
\alias{data_plot_runsequence}
\title{RunSequence plot}
\usage{
data_plot_runsequence(
  data,
  qc_type_subet,
  show_batches = TRUE,
  show_qc_dataset_only = FALSE,
  segment_width = 0.5,
  batches_as_shades = FALSE,
  batch_line_color = "darkred",
  batch_shading_color = "grey60",
  scale_factor = 1,
  show_outlier = TRUE,
  base_size = 12,
  factor_a = NA,
  factor_b = NA
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_type_subet}{Select QC types to be show, NA will select all available QC/Sample types}

\item{show_batches}{Show batches}

\item{show_qc_dataset_only}{Show only available QC types}

\item{segment_width}{Linewidth of the segments}

\item{batches_as_shades}{Show batches as shades}

\item{batch_line_color}{batch separator color}

\item{batch_shading_color}{batch shade color}

\item{scale_factor}{Overall plot scale factor}

\item{show_outlier}{Should samples defined as outlier as separate row}

\item{base_size}{base font size}

\item{factor_a}{First factor referring to a field in the study sample metadata used to show randomization}

\item{factor_b}{First factor referring to a field in the study sample metadata used to show randomization}
}
\value{
ggplot object
}
\description{
RunSequence plot
}
