% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-managment.R
\name{data_set_intensity_var}
\alias{data_set_intensity_var}
\title{Set default variable to be used as feature raw signal value}
\usage{
data_set_intensity_var(
  data = NULL,
  variable_name,
  auto_select = FALSE,
  warnings = TRUE,
  ...
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable_name}{Feature variable to be used as default feature intensity for downstream processing.}

\item{auto_select}{If \code{TRUE} then the first available of these will be used as default: "feature_intensity", "feature_response", "feature_area", "feature_height".}

\item{warnings}{Suppress warnings}

\item{...}{Feature variables to best search for one-by-one when \code{auto-detect = TRUE}}
}
\value{
MidarExperiment object
}
\description{
Sets the raw signal variable used for calculations starting from raw signal
values (i.e., normalization)
}
