% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summarize.R
\name{data_sum_features}
\alias{data_sum_features}
\title{Sum up feature intensities per analyte}
\usage{
data_sum_features(data, qualifier_action = "separate", feature_classes = NA)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qualifier_action}{Character. How to handle qualifier features. To sum them up separately select "separate",}

\item{feature_classes}{Select feature classes to sum up. Default (NA) is to sum up all feature classes according to their analyte_id.
to include them in the sum if quantifier select "include", to not sum them up select "exclude".}
}
\value{
MidarExperiment object
}
\description{
Retrieves analysis IDs of data outliers based  the principal components PCA
with SD or MAD fences
}
