% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midar-batchcorr.R
\name{drift_corr_loess}
\alias{drift_corr_loess}
\title{Performs drift correction based on LOESS smoothing}
\usage{
drift_corr_loess(
  data,
  qc_types,
  smooth_by_batch = TRUE,
  log2_transform = TRUE,
  loess_span = 0.75,
  limit_to_species = NULL,
  apply_only_if_smpl_cv_lower = FALSE,
  treshold_smpl_cv_delta = 0
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types used for drift correction}

\item{smooth_by_batch}{Correct each batch separately (Default is TRUE)}

\item{log2_transform}{log2 transform data during correction (Default is TRUE)}

\item{loess_span}{Loess span width (default is 0.75)}

\item{limit_to_species}{Apply correction only to species matching (RegEx)}

\item{apply_only_if_smpl_cv_lower}{Apply correction only if sample CV decreases (default is FALSE)}

\item{treshold_smpl_cv_delta}{Maximum sample CV change for correction to be applied}
}
\description{
Performs drift correction based on LOESS smoothing
}
