% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-managment.R
\name{features_exclude}
\alias{features_exclude}
\title{Exclude features from the dataset}
\usage{
features_exclude(data, features_exlude, overwrite)
}
\arguments{
\item{data}{MidarExperiment object}

\item{features_exlude}{Vector of feature IDs to exclude from the dataset}

\item{overwrite}{If \code{TRUE} then existing valid_feature flags will be overwritten, otherwise appended}
}
\value{
\code{MidarExperiment} object
}
\description{
Exclude features from the dataset
}
