% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-drift-batch.R
\name{fun_gam_smooth}
\alias{fun_gam_smooth}
\title{Generalized Additive Model (GAM) smoothing helper function}
\usage{
fun_gam_smooth(tbl, reference_qc_types, log_transform_internal = TRUE, ...)
}
\arguments{
\item{tbl}{Table (\code{tibble} or \code{data.frame}) containing the fields \code{qc_type}, \code{x} (run order number), and \code{y} (variable)}

\item{reference_qc_types}{QC types used for the smoothing (fit) by GAM}

\item{log_transform_internal}{Apply log transformation internally for smoothing if \code{TRUE} (default). This does not affect the final data, which remains untransformed.}

\item{...}{Additional arguments forwarded to \code{mgcv::gam}}
}
\value{
List with a \code{data.frame} containing original x and the smoothed y values, and a \code{boolean} value indicating whether the fit failed or not.
}
\description{
Function for penalized spline-based smoothing using GAM, for use by \code{fun_correct_drift}
}
