% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-drift-batch.R
\name{fun_loess}
\alias{fun_loess}
\title{Loess smoothing helper function}
\usage{
fun_loess(tbl, ref_qc_types, log_transform_internal, ...)
}
\arguments{
\item{tbl}{Table (\code{tibble} or \code{data.frame}) containing the fields \code{qc_type}, \code{x} (run order number), and \code{y} (variable)}

\item{ref_qc_types}{QC types used for the smoothing (fit) by loess}

\item{log_transform_internal}{Apply log transformation internally for smoothing if \code{TRUE} (default). This does not affect the final data, which remains untransformed.}

\item{...}{Additional arguments forwarded to Loess}
}
\value{
List with a \code{data.frame} containing original x and the smoothed y values, and a \code{boolean} value indicting whether the fit failed or not not.
}
\description{
Function for loess-based smoothing, for use by \code{fun_correct_drift}
}
