% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-calibrations.R
\name{get_calibration_metrics}
\alias{get_calibration_metrics}
\title{Get Calibration Metrics}
\usage{
get_calibration_metrics(
  data = NULL,
  with_lod = TRUE,
  with_loq = TRUE,
  with_bias = TRUE,
  with_coefficients = TRUE,
  with_sigma = TRUE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object with QC metrics.}

\item{with_lod}{Whether to include LoD in output. Default is \code{TRUE}.}

\item{with_loq}{Whether to include LoQ in output. Default is \code{TRUE}.}

\item{with_bias}{Whether to include bias in output. Default is \code{TRUE}.}

\item{with_coefficients}{Whether to include regression coefficients. Default is \code{TRUE}.}

\item{with_sigma}{Whether to include sigma in output. Default is \code{TRUE}.}
}
\value{
A tibble with exported calibration metrics.
}
\description{
Extracts calibration fit metrics from a \code{MidarExperiment} object.
}
\details{
Requires prior computation of regression results using \code{\link[=calc_calibration_results]{calc_calibration_results()}}.
See its documentation for details.
\subsection{Returned Details and Metrics}{
\itemize{
\item \code{feature_id}: Feature identifier.
\item \code{is_quantifier}: Logical, indicates if the feature is a quantifier.
\item \code{fit_model}: Regression model used for fitting.
\item \code{weighting}: Weighting method used in fitting.
\item \code{lowest_cal}: Lowest nonzero calibration concentration.
\item \code{highest_cal}: Highest  calibration concentration.
\item \code{r.squared}: R-squared value, indicating goodness of fit.
\item \code{coef_a}: Intercept of the regression line
\item \code{coef_b}: Slope of the regression line in \strong{linear} models, or coefficient of the linear term (\code{x}) in \strong{quadratic} models.
\item \code{coef_c}: Coefficient of the quadratic term (\code{x^2}) in \strong{quadratic} models. Returns \code{NA} for \strong{linear} models.
\item \code{sigma}: Standard deviation of residuals.
\item \code{reg_failed}: \code{TRUE} if regression fitting failed.
\item \code{LoD} = 3.3× the sample standard error of residuals / slope of the regression (see Notes).
\item \code{LoQ} = 10× the sample standard error of residuals / slope of the regression (see Notes).
}

\strong{Note:} For LoD/LoQ calculation using \strong{quadratic} fits, the slope used
in the formula is calculated at the lowest nonzero calibration point.
}
}
