% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{get_calibration_report}
\alias{get_calibration_report}
\title{@export}
\usage{
get_calibration_report(
  data,
  qc_types,
  with_lod = TRUE,
  with_loq = TRUE,
  with_bias = TRUE,
  with_coefficients = TRUE,
  with_sigma = TRUE
)
}
\description{
@export
}
