% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{get_calibration_results}
\alias{get_calibration_results}
\title{Get a calibration regression results}
\usage{
get_calibration_results(
  data,
  qc_types,
  with_lod = TRUE,
  with_loq = TRUE,
  with_bias = TRUE,
  with_coefficients = TRUE,
  with_sigma = TRUE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types to be included in the results, aside of CAL. Default is all QC types.}

\item{with_lod}{Include the limit of detection (LoD) in the results}

\item{with_loq}{Include the limit of quantification (LoQ) in the results}

\item{with_bias}{Include the bias in the results}

\item{with_coefficients}{Include the regression coefficients in the results}

\item{with_sigma}{Include the residual standard error in the results}
}
\description{
Get a calibration regression results
}
