% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-math.R
\name{get_mad_tails}
\alias{get_mad_tails}
\title{Get MAD-based tails}
\usage{
get_mad_tails(x, k, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{k}{multiplier for the MAD}

\item{na.rm}{if TRUE then NA values are stripped from x before computation takes place.}
}
\value{
a numeric vector of length 2 with the lower and upper boundaries.
}
\description{
Computes the lower and upper boundaries based on Median Absolute Deviation
(MAD) from a numeric vector. Returns c(NA_real_, NA_real_) if the input vector
has less than 2 elements.
}
\examples{
x <- c(-100,1, 2, 3, 4, 100)
k <- 1.5
get_mad_tails(x, k)
}
