% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-stats.R
\name{get_stats}
\alias{get_stats}
\title{Get FC and p values from a t test contrast}
\usage{
get_stats(
  data,
  feature,
  grouping,
  group_case,
  group_ref,
  paired,
  min_groupsize = 3,
  log_transform = FALSE
)
}
\arguments{
\item{data}{data.frame or tibble}

\item{feature}{feature}

\item{grouping}{grouping}

\item{group_case}{case group}

\item{group_ref}{reference group}

\item{paired}{paired data}

\item{min_groupsize}{min group size for t test}

\item{log_transform}{log transform before t test}
}
\value{
tibble with p values, fdr and log2FC
}
\description{
Get FC and p values from a t test contrast
}
