% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{import_data_csv}
\alias{import_data_csv}
\title{(Depreciated) Import Wide CSV Files}
\usage{
import_data_csv(
  data = NULL,
  path,
  variable_name,
  analysis_id_col = NA,
  import_metadata = TRUE,
  first_feature_column = NA,
  na_strings = "NA"
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{One or more file names with path, or a folder path, which case all *.csv files in this folder will be read.}

\item{variable_name}{Variable type representing the values in the table. Must be one of "intensity", "norm_intensity", "conc", "area", "height", "response")}

\item{analysis_id_col}{Column to be used as analysis_id. \code{NA} (default) used 'analysis_id' if present, or the first column if it contains unique values.}

\item{import_metadata}{Import additional metadata columns (e.g. batch ID, sample type) and add to the \code{MidarExperiment} object.
Only following metadata column names are supported: "qc_type", "batch_id", "is_quantifier", "is_istd", "analysis_order"}

\item{first_feature_column}{Column number of the first column representing the feature values}

\item{na_strings}{A character vector of strings which are to be interpreted as NA values. Blank fields are also considered to be missing values.}
}
\value{
MidarExperiment object
}
\description{
(Depreciated) Import Wide CSV Files
}
\details{
This function is deprecated. Please use \code{\link[=import_data_csv_wide]{import_data_csv_wide()}} instead.
}
\examples{
file_path <- system.file("extdata", "plain_wide_dataset.csv", package = "midar")

mexp <- MidarExperiment()

mexp <- import_data_csv(
  data = mexp,
  path = file_path,
 variable_name = "conc",
 import_metadata = TRUE)

print(mexp)

}
