% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-msdata.R
\name{import_data_mrmkit}
\alias{import_data_mrmkit}
\title{Reads MRMkit results}
\usage{
import_data_mrmkit(data, path)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{file path of MRMkit csv output with raw peak areas}
}
\value{
A tibble in the long format
}
\description{
Reads MRMkit results
}
