% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-read.R
\name{import_metadata_analyses}
\alias{import_metadata_analyses}
\title{Import analysis metadata}
\usage{
import_metadata_analyses(
  data = NULL,
  table = NULL,
  path = NULL,
  sheet = NULL,
  ignore_warnings = FALSE,
  excl_unmatched_analyses = FALSE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object}

\item{table}{A data frame or tibble with analysis (sample) metadata. If \code{path} is also provided, an error will be raised.}

\item{path}{A character string specifying the path to a CSV (.csv) or Excel (.xlsx) file. If \code{table} is also provided, an error will be raised.}

\item{sheet}{Defines the sheet name in case an Excel file is provided.}

\item{ignore_warnings}{Ignore warnings from data validation and proceed with importing metadata}

\item{excl_unmatched_analyses}{Exclude analyses (samples) that have no matching metadata}
}
\value{
An updated \code{MidarExperiment} object
}
\description{
Imports analysis metadata (annotation) from a preloaded data frame or tibble via the \code{data} argument, or from data from a file (CSV or Excel) via the \code{path} argument.
The analysis metadata must contain following columns: \code{analysis_id} and \code{qc_type}. Additional analysis metadata columns are described under details below.
}
\examples{
mexp <- MidarExperiment()
file_path = system.file("extdata", "MHQuant_demo.csv", package = "midar")
mexp <- import_data_masshunter(
  data = mexp,
  path = file_path,
  import_metadata = FALSE)

meta_path = system.file("extdata", "MHQuant_demo_metadata_analyses.csv", package = "midar")

mexp <- import_metadata_analyses(
  data = mexp,
  path = meta_path,
  excl_unmatched_analyses = TRUE)

print(mexp)
}
