% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-import.R
\name{import_metadata_from_data}
\alias{import_metadata_from_data}
\title{Retrieve Metadata from Imported Analysis Data}
\usage{
import_metadata_from_data(data = NULL, qc_type_column_name = "qc_type")
}
\arguments{
\item{data}{A \code{MidarExperiment} object}

\item{qc_type_column_name}{Column name in the imported raw data representing
the \code{qc_type}}
}
\value{
An updated \code{MidarExperiment} object
}
\description{
Retrieves available metadata from the imported analysis data and
associates it with the provided MidarExperiment object.
}
