% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-read.R
\name{import_metadata_midarxlm}
\alias{import_metadata_midarxlm}
\title{Import Metadata from MIDAR Template}
\usage{
import_metadata_midarxlm(
  data = NULL,
  path,
  ignore_warnings = FALSE,
  excl_unmatched_analyses = FALSE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object}

\item{path}{File name and path of the MIDAR template (.xlsm/.xlsx) file}

\item{ignore_warnings}{Ignore warnings from data validation and proceed with importing metadata}

\item{excl_unmatched_analyses}{Exclude analyses (samples) that have no matching metadata}
}
\value{
An updated \code{MidarExperiment} object
}
\description{
Imports metadata from a MIDAR template (.xlsm/.xlsx) file and associates it with analysis data.
}
\examples{
mexp <- MidarExperiment()

mexp <- import_data_masshunter(
  data = mexp,
  path = system.file("extdata", "Example_MHQuant_Small.csv", package = "midar"),
  import_metadata = TRUE)

mexp <- import_metadata_midarxlm(
 data = mexp,
 path = system.file("extdata", "Example_Metadata_Small.xlsm", package = "midar"),
 excl_unmatched_analyses = FALSE)

print(mexp)

}
