% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-read.R
\name{import_metadata_qcconcentrations}
\alias{import_metadata_qcconcentrations}
\title{Import calibration curves metadata}
\usage{
import_metadata_qcconcentrations(
  data = NULL,
  table = NULL,
  path = NULL,
  sheet = NULL,
  ignore_warnings = FALSE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object}

\item{table}{A data frame or tibble with calibration curve metadata. If \code{path} is also provided, an error will be raised.}

\item{path}{A character string specifying the path to a CSV (.csv) or Excel (.xlsx) file. If \code{table} is also provided, an error will be raised.}

\item{sheet}{Defines the sheet name in case an Excel file is provided.}

\item{ignore_warnings}{Ignore warnings from data validation and proceed with importing metadata}
}
\value{
An updated \code{MidarExperiment} object
}
\description{
Imports calibration curve metadata (annotation) from a preloaded data frame or tibble via the \code{data} argument, or from data from a file (CSV or Excel) via the \code{path} argument.
The analysis metadata must contain following columns: \code{analysis_id}, \code{curve_id}, \code{feature_id}, \code{concentration}, and \code{concentration_unit}.
}
