% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidomics.R
\name{lipidomics_get_lipid_class_names}
\alias{lipidomics_get_lipid_class_names}
\title{Retrieve lipid name, lipid class and transition from feature names}
\usage{
lipidomics_get_lipid_class_names(
  data,
  use_as_feature_class = "lipid_class",
  add_transition_names = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{use_as_feature_class}{Set feature_class to lipid_class}

\item{add_transition_names}{add transition name and transition group, based on information in square brackets in feature_id}
}
\value{
MidarExperiment object
}
\description{
Retrieve lipid name, lipid class and transition from feature names
}
