% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-read.R
\name{metadata_from_data}
\alias{metadata_from_data}
\title{Retrieve available metadata from imported analysis data}
\usage{
metadata_from_data(data, qc_type_field = "qc_type")
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
MidarExperiment object
}
\description{
Available information will depend on the format of imported raw analysis data. See \code{\link[=rawdata_import_agilent]{rawdata_import_agilent()}} and \code{\link[=rawdata_import_mrmkit]{rawdata_import_mrmkit()}}
}
