% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-read.R
\name{metadata_import_fromdata}
\alias{metadata_import_fromdata}
\title{Retrieve Metadata from Imported Analysis Data}
\usage{
metadata_import_fromdata(data = NULL, qc_type_column_name = "qc_type")
}
\arguments{
\item{data}{MidarExperiment object containing the imported analysis data}

\item{qc_type_column_name}{Column name in the imported raw data representing the \code{qc_type}}
}
\value{
The input MidarExperiment object with the metadata retrieved and associated
}
\description{
Retrieves available metadata from the imported analysis data and associates it with the provided MidarExperiment object.
}
