% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-access.R
\name{metadata_responsecurves,MidarExperiment-method}
\alias{metadata_responsecurves,MidarExperiment-method}
\title{Get response curve metadata}
\usage{
\S4method{metadata_responsecurves}{MidarExperiment}(x)
}
\value{
Tibble
}
\description{
Get curve IDs and sample amount information for response curves
}
