% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-access.R
\name{metadata_responsecurves<-}
\alias{metadata_responsecurves<-}
\title{metadata_responsecurves method}
\usage{
metadata_responsecurves(x) <- value
}
\arguments{
\item{x}{MidarExperiment object}

\item{value}{Table with response curve metadata}
}
\value{
Tibble
}
\description{
Get curve IDs and sample amount information for response curves
}
