% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-access.R
\name{metadata_responsecurves}
\alias{metadata_responsecurves}
\title{metadata_responsecurves method}
\usage{
metadata_responsecurves(x)
}
\arguments{
\item{x}{MidarExperiment object}
}
\value{
Tibble
}
\description{
Get curve IDs and sample amount information for response curves
}
