% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midar-package.R
\docType{package}
\name{midar-package}
\alias{midar}
\alias{midar-package}
\title{midar: Small Molecule Mass Spectrometry Data Post-Processing, Quality Control, Exploration and Reporting}
\description{
R functions to facilitate managing, processing and evaluating lipidomics data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SLINGhub/midar}
  \item Report bugs at \url{https://github.com/SLINGhub/midar/issues}
}

}
\author{
\strong{Maintainer}: Bo Burla \email{bo.burla@nus.edu.sg} (\href{https://orcid.org/0000-0002-5918-3249}{ORCID})

}
\keyword{internal}
