% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{MidarExperiment}
\alias{MidarExperiment}
\title{Constructor for the MidarExperiment object.}
\usage{
MidarExperiment(title = "", analysis_type = NA_character_)
}
\arguments{
\item{title}{Title of experiment}

\item{analysis_type}{Analysis type, one of "lipidomics", "metabolomics", "externalcalib", "others"}
}
\value{
\code{MidarExperiment} object
}
\description{
Constructor for the MidarExperiment object.
}
