% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipidomics.R
\name{parse_lipid_feature_names}
\alias{parse_lipid_feature_names}
\title{Get lipid class, species and transition names}
\usage{
parse_lipid_feature_names(
  tbl,
  use_as_feature_class = "lipid_class_lcb",
  add_transition_names = FALSE,
  add_chain_composition = TRUE
)
}
\arguments{
\item{tbl}{A data frame containing a \code{feature_id} column}

\item{use_as_feature_class}{Set feature_class to lipid_class}

\item{add_transition_names}{add transition name and transition group, based on information in square brackets in feature_id}

\item{add_chain_composition}{add total_c and total_db to the dataset}
}
\value{
MidarExperiment object
}
\description{
This function retrieves lipid class, species and transition names from the \code{feature_id} column and adds them as columns to the dataset.
}
