% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{parse_mrmkit_result}
\alias{parse_mrmkit_result}
\title{Parses MRMkit peak integration results into a tibble}
\usage{
parse_mrmkit_result(path, silent = FALSE)
}
\arguments{
\item{path}{File name of the MRMkit result file (*.tsv or *.csv)}

\item{silent}{No comments printed}
}
\value{
A tibble in the long format
}
\description{
Parses MRMkit peak integration results into a tibble
}
\examples{

file_path = system.file("extdata", "MRMkit_demo.tsv", package = "midar")

tbl <- parse_mrmkit_result(path = file_path)

head(tbl)
}
