% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{parse_plain_wide_csv}
\alias{parse_plain_wide_csv}
\title{Parses a plain wide CSV file}
\usage{
parse_plain_wide_csv(
  path,
  variable_name,
  analysis_id_col = NA,
  import_metadata = TRUE,
  first_feature_column = NA,
  na_strings = "NA"
)
}
\arguments{
\item{path}{path name and path of a plain long-format CSV file}

\item{variable_name}{Name of the variable representing the values in the table. Must be one of "intensity", "norm_intensity", "conc", "area", "height", "response")}

\item{analysis_id_col}{Column to be used as analysis_id}

\item{import_metadata}{Import additional metadata columns (e.g. batch ID, sample type) and add to the \code{MidarExperiment} object}

\item{first_feature_column}{Column number of the first column representing the feature values}

\item{na_strings}{A character vector of strings which are to be interpreted as NA values. Blank fields are also considered to be missing values.}
}
\value{
A tibble in the long format
}
\description{
Parses a CSV table with analysis/samples in rows, features values in columns.
}
\examples{
file_path <- system.file("extdata", "plain_wide_dataset.csv", package = "midar")

tbl <- parse_plain_wide_csv(
 path = file_path,
 variable_name = "conc",
 analysis_id_col = "analysis_id",
 import_metadata = TRUE)

head(tbl)
}
