% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-calibcurves.R
\name{plot_calibrationcurves}
\alias{plot_calibrationcurves}
\title{Plots calibration curves for each measured feature.}
\usage{
plot_calibrationcurves(
  data = NULL,
  variable = "intensity",
  overwrite_metadata = FALSE,
  fit_method = c("linear", "quadratic"),
  fit_weighting = c(NA, "none", "1/x", "1/x^2"),
  log_axes = FALSE,
  filter_data = FALSE,
  output_pdf = FALSE,
  include_feature_filter = "",
  exclude_feature_filter = "",
  max_regression_value = NA,
  path = "",
  rows_page = 4,
  cols_page = 5,
  specific_page = NA,
  point_size = 1.5,
  line_width = 0.7,
  plot_scaling_factor = 1,
  page_orientation = "LANDSCAPE",
  font_base_size = 7,
  show_progress = TRUE,
  return_plots = FALSE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object .}

\item{variable}{The variable name to plot on the y-axis, usually a measure of intensity. Defaults to \code{"intensity"}}

\item{overwrite_metadata}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the function will ignore any fit method and weighting settings defined in the metadata and use the provided \code{fit_method} and \code{fit_weighting} values for all analytes.}

\item{fit_method}{A character string specifying the default regression fit method to use for the calibration curve. Must be one of \code{"linear"} or \code{"quadratic"}. This method will be applied if no specific fit method is defined for a feature in the metadata.}

\item{fit_weighting}{A character string specifying the default weighting method for the regression points in the calibration curve. Must be one of \code{"none"}, \code{"1/x"}, or \code{"1/x^2"}. This method will be applied if no specific weighting method is defined for a feature in the metadata.}

\item{log_axes}{Logical. If \code{TRUE}, the x and y axes are displayed in logarithmic scale. Defaults to \code{FALSE}.}

\item{filter_data}{Logical. Indicates whether to use quality control (QC) filtered data (\code{TRUE}) or the raw data (\code{FALSE}). Defaults to \code{FALSE}.}

\item{output_pdf}{Logical. If \code{TRUE}, saves the generated plots as a PDF file. Defaults to \code{FALSE}.}

\item{include_feature_filter}{A regex pattern to filter and include features that match the criteria. If omitted, all features are considered.}

\item{exclude_feature_filter}{A regex pattern to filter and exclude features that match the criteria. If omitted, no features are excluded.}

\item{max_regression_value}{The maximum x value (analyzed amount) for which the regression line is fitted. Defaults to \code{NA}, considering all data points.}

\item{path}{The file path for saving the PDF. Defaults to an empty string, meaning no file is saved.}

\item{rows_page}{The number of rows of plots per page for the PDF output. Defaults to 4.}

\item{cols_page}{The number of columns of plots per page for the PDF output. Defaults to 5.}

\item{specific_page}{An integer specifying a particular page to plot. If \code{NA}, all pages are plotted.}

\item{point_size}{Numeric value specifying the size of points in the plot. Defaults to 1.5.}

\item{line_width}{Numeric value specifying the width of regression lines. Defaults to 0.7.}

\item{plot_scaling_factor}{A universal scaling factor for fonts, symbols, and lines. Defaults to 1.}

\item{page_orientation}{The orientation of the PDF paper, either \code{"LANDSCAPE"} or \code{"PORTRAIT"}. Defaults to \code{"LANDSCAPE"}.}

\item{font_base_size}{The base font size for text in the plots. Defaults to 7.}

\item{show_progress}{Logical. Show a progress bar during plot creation if set to \code{TRUE}. Defaults to \code{TRUE}.}

\item{return_plots}{Logical. If \code{TRUE}, returns the plots as a list of \code{ggplot2} objects. Each item represents a page of plots. Defaults to \code{FALSE}.}
}
\value{
If \code{return_plots} is \code{TRUE}, a list of \code{ggplot2} objects is returned. Otherwise, the function may return a plot output and/or save a PDF, depending on the \code{output_pdf} parameter.
}
\description{
Plots calibration curves for each measured feature.
}
