% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-cv.R
\name{plot_normalization_qc}
\alias{plot_normalization_qc}
\title{Compare \%CV values before and after normalization}
\usage{
plot_normalization_qc(
  data = NULL,
  before_norm_var = c("intensity", "norm_intensity"),
  after_norm_var = c("norm_intensity", "conc"),
  qc_type,
  facet_by_class = FALSE,
  filter_data = FALSE,
  include_qualifier = FALSE,
  cv_threshold_value = 25,
  xlim = c(0, NA),
  ylim = c(0, NA),
  ncol = 5,
  point_size = 1,
  point_alpha = 0.5,
  font_base_size = 8
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object}

\item{before_norm_var}{A string specifying the variable from the QC metrics
table to be used for the x-axis (before normalization).}

\item{after_norm_var}{A string specifying the variable from the QC metrics
table to be used for the y-axis (after normalization).}

\item{qc_type}{QC type to be used for the comparison (one of: "SPL", "BQC",
"TQC", "NIST", "LTR").}

\item{facet_by_class}{If \code{TRUE}, facets the plot by \code{feature_class}, as defined
in the feature metadata.}

\item{filter_data}{Whether to use all data (default) or only
QC-filtered data (filtered via \code{\link[=filter_features_qc]{filter_features_qc()}}).}

\item{include_qualifier}{Whether to include qualifier features
(default is \code{TRUE}).}

\item{cv_threshold_value}{Numerical threshold value to be shown as dashed
lines in the plot (default is \code{25}).}

\item{xlim}{Numeric vector of length 2 for x-axis limits. Use \code{NA} for
auto-scaling (default is \code{c(0, NA)}).}

\item{ylim}{Numeric vector of length 2 for y-axis limits. Use \code{NA} for
auto-scaling (default is \code{c(0, NA)}).}

\item{ncol}{Number of facet columns per page, representing
different feature classes (default is \code{5}). Only used if
\code{facet_by_class = TRUE}.}

\item{point_size}{Size of points in millimeters (default is \code{1}).}

\item{point_alpha}{Transparency of points (default is \code{0.5}).}

\item{font_base_size}{Base font size in points (default is \code{8}).}

\item{equality_line}{Whether to show a line indicating
identical values in both compared variables (default is \code{FALSE})..}
}
\value{
A \code{ggplot2} object representing the scatter plot comparing CV values
before and after normalization.
}
\description{
This function compares the coefficient of variation (CV) of QC or study
samples before and after normalization. It preselects the relevant QC metrics
based on the chosen arguments and visualizes the comparison through a scatter
plot. The plot can be faceted by \code{feature_class}.
}
\details{
The function preselects the corresponding variables from the QC metrics and uses
\code{\link[=plot_qcmetrics_comparison]{plot_qcmetrics_comparison()}} to visualize the results.
\itemize{
\item The data must be normalized before using \code{\link[=normalize_features]{normalize_features()}} followed by
calculation of the QC metrics table via \code{\link[=calc_qc_metrics]{calc_qc_metrics()}} or
\code{\link[=filter_features_qc]{filter_features_qc()}}, see examples below.
\item When \code{facet_by_class = TRUE}, then the \code{feature_class} must be defined in the
metadata or retrieved via specific functions, e.g., \code{\link[=get_lipid_class_names]{get_lipid_class_names()}}.
}
}
\examples{
# Example usage:
mexp <- lipidomics_dataset
mexp <- normalize_by_istd(mexp)
mexp <- calc_qc_metrics(mexp)
plot_normalization_qc(
  data = mexp,
  before_norm_var = "intensity",
  after_norm_var = "norm_intensity",
  qc_type = "SPL",
  filter_data = TRUE,
  facet_by_class = TRUE,
  cv_threshold_value = 25
)

}
\seealso{
\code{\link[=plot_qcmetrics_comparison]{plot_qcmetrics_comparison()}}, \code{\link[=calc_qc_metrics]{calc_qc_metrics()}}, \code{\link[=filter_features_qc]{filter_features_qc()}},
}
