% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_QC.R
\name{plot_pca_qc}
\alias{plot_pca_qc}
\title{PCA plot for QC}
\usage{
plot_pca_qc(
  data,
  variable,
  dim_x,
  dim_y,
  log_transform,
  point_size = 2,
  fill_alpha = 0.1
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable}{which variable to use for plot}

\item{dim_x}{PCA dimension on x axis}

\item{dim_y}{PCA dimension on y axis}

\item{log_transform}{log transform data for plot}

\item{point_size}{size of points}

\item{fill_alpha}{transparency of points}
}
\value{
ggplot2 object
}
\description{
PCA plot for QC
}
