% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-qc-matrixeffects.R
\name{plot_qc_matrixeffects}
\alias{plot_qc_matrixeffects}
\title{Plot standardized feature intensities grouped by QC type}
\usage{
plot_qc_matrixeffects(
  data,
  variable = "intensity",
  qc_types = c("SPL", "TQC", "PBLK", "BQC"),
  batchwise_normalization = TRUE,
  include_qualifier = FALSE,
  only_istd = TRUE,
  include_feature_filter = NA,
  exclude_feature_filter = NA,
  min_median_value = NA,
  ylim = c(-NA, NA),
  point_size = 0.5,
  dodge_width = 0.6,
  point_alpha = 0.3,
  box_alpha = 0.3,
  box_linewidth = 0.5,
  font_base_size = 8,
  angle_x = 45
)
}
\arguments{
\item{data}{A MidarExperiment object}

\item{variable}{A character string indicating the variable to use for PCA
analysis. Must be one of: "area", "height", "intensity", "norm_intensity", "response",
"conc", "conc_raw", "rt", "fwhm".}

\item{qc_types}{A character vector specifying the QC types to plot. It
must contain at least one element. The default is \code{NA}, which means any
of the non-blank QC types ("SPL", "TQC", "BQC", "HQC", "MQC", "LQC",
"NIST", "LTR") will be plotted if present in the dataset.}

\item{batchwise_normalization}{A logical value indicating whether to normalize the signals by batch instead of globally.}

\item{include_qualifier}{A logical value indicating whether to include
qualifier features. Default is \code{TRUE}.}

\item{only_istd}{A logical value indicating whether to include features used as internal
standards (ISTD).  Default is \code{TRUE}. Set to \code{FALSE} in combination with feature_filter parameters to show other features.}

\item{include_feature_filter}{A character or regex pattern used to filter
features by \code{feature_id}. If \code{NA} or an empty string (\code{""}) is provided,
the filter is ignored. When a vector of length > 1 is supplied, only
features with exactly these names are selected (applied individually as
OR conditions).}

\item{exclude_feature_filter}{A character or regex pattern used to exclude
features by \code{feature_id}. If \code{NA} or an empty string (\code{""}) is provided,
the filter is ignored. When a vector of length > 1 is supplied, only
features with exactly these names are excluded (applied individually as
OR conditions).}

\item{min_median_value}{Minimum median
feature value (as determined by the \code{variable}) across all samples from
selected QC types that must be met for a feature to be included in the
PCA analysis. \code{NA} (default) means no filtering will be applied. This
parameter provides an fast way to exclude noisy features from the
analysis. However, it is recommended to use \code{filter_data} with
\code{\link[=filter_features_qc]{filter_features_qc()}}.}

\item{ylim}{A numeric vector of length 2 specifying the y-axis limits.}

\item{point_size}{A numeric value indicating the size of points in
millimeters. Default is 2.}

\item{dodge_width}{Numeric. Width used to dodge overlapping points by \code{qc_type}. Default is \code{0.6}.}

\item{point_alpha}{Numeric. Transparency of the plotted points. Default is \code{0.7}.}

\item{box_alpha}{Numeric. Transparency of the boxplot. Default is \code{0.3}.}

\item{box_linewidth}{Numeric. Width of the boxplot lines. Default is \code{0.5}.}

\item{font_base_size}{A numeric value indicating the base font size for
plot text elements. Default is 8.}

\item{angle_x}{Numeric. Angle of the x-axis text labels. Default is \code{45}.}
}
\value{
A \code{ggplot} object showing the grouped standardized beeswarm plot.
}
\description{
This function creates a grouped beeswarm plot of standardized feature intensities,
where the y-axis represents intensity standardized such that the mean across all
features is 100\%. Points are grouped by \code{qc_type} and spread using quasirandom jitter.
}
