% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-filtering.R
\name{plot_qc_summary_overall}
\alias{plot_qc_summary_overall}
\title{Plot Overall QC Filtering Summary}
\usage{
plot_qc_summary_overall(data = NULL, with_venn = TRUE, font_base_size = 8)
}
\arguments{
\item{data}{MidarExperiment object}

\item{with_venn}{Whether to include a Venn diagram summarizing the features excluded due to different QC criteria. Default is \code{TRUE}.}

\item{font_base_size}{The base font size for the plot. Default is \code{8}.}
}
\value{
A \code{ggplot2} object showing the feature QC filtering summary with or without a Venn diagram.
}
\description{
This function generates a summary of the feature QC filtering process, visualizing the number of features that passed or failed the various QC criteria.
It includes a Venn diagram showing the features excluded due to different filtering criteria such as signal-to-blank ratios, CV thresholds, and linearity.
The criteria are applied hierarchically, meaning a feature must pass all lower-tier filters before being considered for failure on higher-tier filters. See \code{\link[=plot_qc_summary_byclass]{plot_qc_summary_byclass()}} for more information.
}
\details{
The QC filtering process follows a hierarchical structure, where features are first evaluated against lower-level filters such as signal-to-blank ratios and limit of detection (LOD).
Only features that pass these basic criteria are then subjected to higher-level filters like the coefficient of variation (CV) or linear regression results.
A feature will only fail a higher-level filter (such as \code{CV} or \code{R-squared}) if it has passed all previous lower-level filters.
This ensures that features are evaluated progressively, starting from fundamental quality checks up to more stringent filtering criteria.
}
