% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_QC.R
\name{plot_qc_summary_venn}
\alias{plot_qc_summary_venn}
\title{Plot summary of QC filtering}
\usage{
plot_qc_summary_venn(data, user_defined_keeper, base_size)
}
\arguments{
\item{data}{MidarExperiment object}

\item{user_defined_keeper}{Include user-defined feature inclusion list, even they did not pass QC filtering}

\item{base_size}{font size of plots}
}
\value{
ggplot2 object
}
\description{
Plot summary of QC filtering
}
