% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-responsecurves.R
\name{plot_responsecurves}
\alias{plot_responsecurves}
\title{Plot Response Curves}
\usage{
plot_responsecurves(
  data = NULL,
  variable = "intensity",
  filter_data = FALSE,
  qc_types = NA,
  include_qualifier = TRUE,
  include_istd = TRUE,
  include_feature_filter = NA,
  exclude_feature_filter = NA,
  max_regression_value = NA,
  output_pdf = FALSE,
  path = NA,
  return_plots = FALSE,
  color_curves = c("#4575b4", "#91bfdb", "#fc8d59", "#d73027"),
  point_size = 1.5,
  line_width = 0.7,
  font_base_size = 7,
  rows_page = 4,
  cols_page = 5,
  specific_page = NA,
  page_orientation = "LANDSCAPE",
  show_progress = TRUE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object containing the dataset and metadata.}

\item{variable}{The variable to plot on the y-axis.}

\item{filter_data}{Whether to use all data (default) or only
QC-filtered data (filtered via \code{\link[=filter_features_qc]{filter_features_qc()}}).}

\item{include_qualifier}{Logical, whether to include qualifier features. Default is \code{TRUE}.}

\item{include_istd}{Logical, whether to include internal standard (ISTD) features. Default is \code{TRUE}.}

\item{include_feature_filter}{A regex pattern or a vector of feature names used to filter features by \code{feature_id}.
If \code{NA} or an empty string (\code{""}) is provided, the filter is ignored. When a vector of length > 1 is supplied,
is supplied, only features with exactly these names are selected (applied individually as OR conditions).}

\item{exclude_feature_filter}{A regex pattern or a vector of feature names to exclude features by feature_id.
If \code{NA} or an empty string (\code{""}) is provided, the filter is ignored. When a vector of length > 1 is supplied,
is supplied, only features with exactly these names are excluded (applied individually as OR conditions).}

\item{max_regression_value}{The maximum sample_amount (x) value for fitting
the regression line. If \code{NA}, regression is based on all data points.}

\item{output_pdf}{If \code{TRUE}, saves the generated plots as a PDF
file. Wjen \code{FALSE}, plots are directly plotted.}

\item{path}{The file path for saving the PDF. Must be defined if
\code{output_pdf} is \code{TRUE}.}

\item{return_plots}{Logical. If \code{TRUE}, returns the plots as a list of
\code{ggplot2} objects.}

\item{color_curves}{useA vector of colors for the curves.}

\item{point_size}{Size of points in millimeters.}

\item{line_width}{Width of regression lines.}

\item{font_base_size}{Base font size for text.}

\item{rows_page}{Number of rows of plots per page.}

\item{cols_page}{Number of columns of plots per page.}

\item{specific_page}{An integer specifying a specific page to plot. If
\code{NA} (default), all pages are plotted.}

\item{page_orientation}{Orientation of the PDF paper: \code{"LANDSCAPE"} or
\code{"PORTRAIT"}.}

\item{show_progress}{Logical. If \code{TRUE}, displays a progress bar during
plot creation.}
}
\value{
If \code{return_plots} is \code{TRUE}, a list of \code{ggplot2} objects is
returned. Otherwise, the function saves the plot output or does not return
anything.
}
\description{
Generates response curve plots for each measured feature. Plots can be
saved as multi-page PDFs or directly displayed. Multiple response curves,
each with a linear regression line, can be plotted on the same graph.
Each feature is displayed as a separate facet.
}
