% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_QC.R
\name{plot_runboxplots}
\alias{plot_runboxplots}
\title{Sample sequence Boxplots}
\usage{
plot_runboxplots(
  data,
  relative_log_abundances,
  plot_variable,
  use_qc_filtered_data,
  min_feature_intensity,
  qc_types,
  ignore_outliers,
  feature_filter,
  show_batches,
  batches_as_shades,
  batch_line_color,
  batch_shading_color,
  base_size = 8
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{relative_log_abundances}{Plot as RLA (Relative Log Abundnance Plot)}

\item{plot_variable}{Variable to plot}

\item{use_qc_filtered_data}{Use QC-filtered data}

\item{min_feature_intensity}{Exclude features with median signal below this value}

\item{qc_types}{QC types to be plotted}

\item{ignore_outliers}{Exclude outlier values based on 3x IQR fences}

\item{feature_filter}{Filter text to select specific features (regex string)}

\item{show_batches}{Show batches}

\item{batches_as_shades}{Show batches as shades}

\item{batch_line_color}{batch separator color}

\item{batch_shading_color}{batch shade color}

\item{base_size}{base font size for plots (default is 8)}
}
\value{
ggplot object
}
\description{
Sample sequence Boxplots
}
