% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_QC.R
\name{plot_runscatter}
\alias{plot_runscatter}
\title{RunScatter plot}
\usage{
plot_runscatter(
  data,
  plot_variable = c("feature_intensity", "feature_norm_intensity", "feature_conc"),
  feature_filter = "",
  filter_exclude = FALSE,
  cap_outliers = FALSE,
  cap_qc_iqr_factor = 1.5,
  cap_spl_iqr_factor = 1.5,
  cap_top_n_values = NA,
  qc_type_fit = "BQC",
  show_driftcorrection = FALSE,
  show_trend_samples,
  trend_samples_fun = "loess",
  trend_samples_col = "",
  after_correction = FALSE,
  plot_other_qc = TRUE,
  show_batches = FALSE,
  batches_as_shades = TRUE,
  batch_line_color = "#9dbecf",
  batch_shading_color = "grey90",
  outputPDF = FALSE,
  filename = "",
  cols_page = 4,
  rows_page = 3,
  annot_scale = 1,
  paper_orientation = "LANDSCAPE",
  point_transparency = 1,
  point_size = 2,
  point_stroke_width = 0.8,
  page_no = NA,
  y_label_text = NA,
  silent = FALSE,
  return_plot_list = FALSE,
  base_size = 7,
  show_gridlines = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{plot_variable}{Variable to plot}

\item{feature_filter}{Filter features containing}

\item{filter_exclude}{Exclude or include feature_filter}

\item{cap_outliers}{Cap outliers to a specific range in the plot}

\item{cap_qc_iqr_factor}{Multiplicator for the upper Tukey's IQR fence use for outlier capping of QCs}

\item{cap_spl_iqr_factor}{Multiplicator for the upper Tukey's IQR fence use for outlier capping of samples}

\item{cap_top_n_values}{Cap top n values}

\item{qc_type_fit}{QC TYPE used for loess fit}

\item{show_driftcorrection}{Show drift correction}

\item{show_trend_samples}{Fit trend line for study samples, if show_driftcorrection is TRUE}

\item{trend_samples_fun}{Function used for drift correction. Default 'loess'}

\item{trend_samples_col}{Color of drift line}

\item{after_correction}{Show before/after correction}

\item{plot_other_qc}{Plot all QCS}

\item{show_batches}{Show batches}

\item{batches_as_shades}{Show batches as shades}

\item{batch_line_color}{batch separator color}

\item{batch_shading_color}{batch shade color}

\item{outputPDF}{save as PDF}

\item{filename}{file name of PDF}

\item{cols_page}{columns per page}

\item{rows_page}{rows per page}

\item{annot_scale}{scale factor of text elements}

\item{paper_orientation}{Landscape/Portrait}

\item{point_transparency}{Alpha of points}

\item{point_size}{point size}

\item{point_stroke_width}{point stroke width}

\item{page_no}{Show page number}

\item{y_label_text}{Overwrite y label with this text}

\item{silent}{Verbose or silent}

\item{return_plot_list}{return list with plots}

\item{base_size}{base font size of plot}

\item{show_gridlines}{show x and y major gridlines}
}
\value{
A list of ggplot2 plots or NULL
}
\description{
RunScatter plot
}
