% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-runorder.R
\name{plot_runsequence}
\alias{plot_runsequence}
\title{RunSequence Plot}
\usage{
plot_runsequence(
  data = NULL,
  qc_types = NA,
  show_batches = TRUE,
  show_timestamp = FALSE,
  add_info_title = TRUE,
  single_row = FALSE,
  segment_width = 0.5,
  batch_zebra_stripe = FALSE,
  batch_line_color = "#b6f0c5",
  batch_fill_color = "grey90",
  base_font_size = 8
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types to be plotted. Can be a vector of QC types or a
regular expression pattern. \code{NA} (default) displays all available QC/Sample
types.}

\item{show_batches}{Logical, whether to show batch separators in the plot}

\item{show_timestamp}{Logical, whether to use the acquisition timestamp as
the x-axis instead of the run sequence number}

\item{add_info_title}{Logical, whether to add a title with the experiment
title, analysis date, and analysis times}

\item{single_row}{Logical, whether to show all QC types in a single row}

\item{batch_zebra_stripe}{Logical, whether to show batches as shaded areas
instead of line separators}

\item{batch_line_color}{Color of the batch separator lines}

\item{batch_fill_color}{Color of the batch shaded areas}

\item{base_font_size}{Numeric, base font size for the plot}
}
\value{
A ggplot object representing the run sequence plot
}
\description{
The RunSequence plot provides an overview of the analysis design and
timelines, which can be useful for subsequent processing steps. The plot
illustrates the batch structure, the quality control (QC) samples included
with their respective positions, and additional information regarding the
date, duration, and run time of the analysis.

Setting \code{show_timestamp = TRUE} allows you to check for any interruptions in
the analysis timeline.
}
