% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_qc.R
\name{plot_x_vs_y}
\alias{plot_x_vs_y}
\title{Plot comparison of analytical CV before and after ISTD-normalization}
\usage{
plot_x_vs_y(
  data,
  x,
  y,
  only_quantifier = TRUE,
  xlim = c(0, NA),
  ylim = c(0, NA),
  ncol = 5,
  scale_factor = 1,
  point_size = 3,
  with_histogram = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{x}{Variable of QC metrics table used for x axis}

\item{y}{Variable of QC metrics table used for y axis}

\item{only_quantifier}{Show only quantifier}

\item{xlim}{Lower and upper limits of the x axis as vector}

\item{ylim}{Lower and upper limits of the y axis as vector}

\item{ncol}{Numer of columns with facets}

\item{scale_factor}{Scale factor for text labels}

\item{point_size}{Point size}

\item{with_histogram}{Show side histograms indicating distribution}
}
\value{
ggplot2 object
}
\description{
Plot comparison of analytical CV before and after ISTD-normalization
}
