% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-filtering.R
\name{qc_apply_feature_filter}
\alias{qc_apply_feature_filter}
\title{Filter dataset according to QC and other criteria}
\usage{
qc_apply_feature_filter(
  data = NULL,
  overwrite = TRUE,
  batch_medians = FALSE,
  qualifier.include = FALSE,
  istd.include = FALSE,
  missing.intensity.spl.prop.max = NA,
  missing.norm_intensity.spl.prop.max = NA,
  missing.conc.spl.prop.max = NA,
  intensity.min.bqc.min = NA,
  intensity.min.tqc.min = NA,
  intensity.min.spl.min = NA,
  intensity.median.bqc.min = NA,
  intensity.median.tqc.min = NA,
  intensity.median.spl.min = NA,
  intensity.max.spl.min = NA,
  signalblank.median.pblk.min = NA,
  signalblank.median.ublk.min = NA,
  signalblank.median.sblk.min = NA,
  cv.conc.bqc.max = NA,
  cv.conc.tqc.max = NA,
  cv.intensity.bqc.min = NA,
  cv.intensity.tqc.min = NA,
  dratio.conc.bqc.sd.max = NA,
  dratio.conc.tqc.sd.max = NA,
  dratio.conc.bqc.mad.max = NA,
  dratio.conc.tqc.mad.max = NA,
  response.curve.id = NA,
  response.curve.summary = NA,
  response.rsquare.min = NA,
  response.slope.min = NA,
  response.slope.max = NA,
  response.yintersect.max = NA,
  outlier.technical.exlude = FALSE,
  features_to_keep = NULL
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{overwrite}{Add/change previously defined qc filters or define a new set of filters. Default is \code{FALSE}.}

\item{batch_medians}{Use median of batch-wise derived QC values. Default is FALSE.}

\item{qualifier.include}{Include qualifier features}

\item{istd.include}{Include Internal Standards (ISTD). If set to FALSE, meaning ISTDs will be included, then \code{min_signal_blank_ratio} is ignored, because the the S/B is based on Processed Blanks (PBLK) that contain ISTDs.}

\item{missing.intensity.spl.prop.max}{NA Proportion of missing raw intensities}

\item{missing.norm_intensity.spl.prop.max}{NA Proportion of missing normalized intensities}

\item{missing.conc.spl.prop.max}{NA Proportion of missing final concentrations}

\item{intensity.min.bqc.min}{Minimum median feature intensity of BQC}

\item{intensity.min.tqc.min}{Minimum median feature intensity of TQC}

\item{intensity.min.spl.min}{Minimum median feature intensity of study samples (SPL)}

\item{intensity.median.bqc.min}{Minimum median feature intensity of BQC}

\item{intensity.median.tqc.min}{Minimum median feature intensity of TQC}

\item{intensity.median.spl.min}{Minimum median feature intensity of study samples (SPL)}

\item{intensity.max.spl.min}{Minimum maximun feature intensity oof study samples (SPL)}

\item{signalblank.median.pblk.min}{= Signal-to-Blank ratio. Calculated from the median of study samples and the median of the Process Blank (PBLK)}

\item{signalblank.median.ublk.min}{= Signal-to-Blank ratio. Calculated from the median of study samples and the median of the Unprocessed Blank (UBLK)}

\item{signalblank.median.sblk.min}{= Signal-to-Blank ratio. Calculated from the median of study samples and the median of the Solvent Blank (SBLK)}

\item{cv.conc.bqc.max}{= Maximum \%CV of BQC}

\item{cv.conc.tqc.max}{Maximum \%CV of TQC}

\item{cv.intensity.bqc.min}{Maximum \%CV of BQC}

\item{cv.intensity.tqc.min}{Maximum \%CV of TQC}

\item{dratio.conc.bqc.sd.max}{D-ratio defined as CV_BQC/CV_SPL, based on sd}

\item{dratio.conc.tqc.sd.max}{D-ratio defined as CV_TQC/CV_SPL, based on sd}

\item{dratio.conc.bqc.mad.max}{D-ratio defined as CV_BQC/CV_SPL, based on mad}

\item{dratio.conc.tqc.mad.max}{D-ratio defined as CV_TQC/CV_SPL, based on mad}

\item{response.curve.id}{A single value or vector specifying the identifiers or indices of the response curves to be used for filtering. This can be either numeric indices or text identifiers.}

\item{response.curve.summary}{A string indicating how to summarize the metrics of multiple selected curves, if applicable. Must be either 'mean', 'median', 'best', or 'worst'.}

\item{response.rsquare.min}{The minimum acceptable R squared value for the response curve(s). Refer to the details section for more information.}

\item{response.slope.min}{The minimum normalized slope for the response curve(s). Refer to the details section for further explanation.}

\item{response.slope.max}{The maximum normalized slope for the response curve(s). Refer to the details section for additional details.}

\item{response.yintersect.max}{The maximum allowable normalized y0 intercept for the response curve(s).}

\item{outlier.technical.exlude}{Remove samples classified as outliers}

\item{features_to_keep}{A vector specifying the features that should be retained, regardless of whether they meet the specified quality control filtering criteria.}
}
\value{
MidarExperiment object
}
\description{
Filter dataset according to QC parameter criteria, remove features that are internal standards (ISTDs) or not annotated as quantifier (optional).
Exclude features and analyses that were annotated as not valid in the metadata (valid_feature, valid_analysis).

Missing qc parameter values, e.g. no mean BQC intensity, because no BQC values present in the data, are handled as following:
\itemize{
\item If a filter is applied (e.g. \code{intensity.min.bqc.min}) but qc value is NA then this feature will be excluded.
\item If a filter is no applied then no matter if qc value is defined or NA, no filtering will be applied
}

Note: When \code{istd.include} is TRUE, then \code{signalblank.median.sblk.min} and \code{signalblank.median.sblk.min} are ignored for ISTDs, because these blanks are defined as containing ISTDs.
}
\details{
Response curves

A normalized slope of 1 indicates a proportional relationship between the response and the sample amount. Negative slopes reflect an inverse relationship, where the response decreases as the sample amount increases.
The normalized y-intercept represents the value of the response when the sample amount is zero. A value of 1 corresponds to the maximum measured response on the curve.
}
