% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-filtering.R
\name{qc_calculate_metrics}
\alias{qc_calculate_metrics}
\title{Calculate QC metrics for each feature}
\usage{
qc_calculate_metrics(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
MidarExperiment object
}
\description{
Calculate QC metrics for each feature
}
