% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-cv.R
\name{qc_plot_normalization_cv}
\alias{qc_plot_normalization_cv}
\title{Comparison of CV values before and after normalization}
\usage{
qc_plot_normalization_cv(
  data = NULL,
  filter_data,
  qc_type,
  var_before = c("intensity"),
  var_after = c("norm_intensity"),
  cv_treshhold = 25,
  only_quantifier = TRUE,
  xlim = c(0, NA),
  ylim = c(0, NA),
  ncol = 5,
  point_size = 1,
  point_alpha = 0.5,
  scale_factor = 1,
  font_base_size = 8
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filter_data}{Use all (default) or qc-filtered data}

\item{qc_type}{QC type to use for plot.}

\item{var_before}{Variable of QC metrics table used for x axis}

\item{var_after}{Variable of QC metrics table used for y axis}

\item{cv_treshhold}{Threshold values shown in plot as dashed lines}

\item{only_quantifier}{Show only quantifier}

\item{xlim}{Lower and upper limits of the x axis as vector}

\item{ylim}{Lower and upper limits of the y axis as vector}

\item{ncol}{Number of facet columns}

\item{point_size}{size of points. Default is \code{2}}

\item{point_alpha}{transparency of points}

\item{scale_factor}{Scale factor for text labels}

\item{font_base_size}{Base font size for plot text elements}
}
\value{
ggplot2 object
}
\description{
Comparison of CV values before and after normalization
}
