% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-pca.R
\name{qc_plot_pca}
\alias{qc_plot_pca}
\title{PCA plot for QC}
\usage{
qc_plot_pca(
  data,
  variable,
  filter_data,
  pca_dim = c(1, 2),
  qc_types = c("SPL", "BQC", "TQC", "NIST", "LTR"),
  label_k_mad = 3,
  log_transform = TRUE,
  remove_istds = TRUE,
  min_median_signal = NA,
  point_size = 2,
  point_alpha = 0.7,
  ellipse_alpha = 0.8,
  font_base_size = 8,
  hide_label_text = NA
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable}{Which variable to use for the PCA. Must be any of "area", "height", "intensity", "response", "conc", "conc_raw", "rt", "fwhm".}

\item{filter_data}{Use all (default) or qc-filtered data}

\item{pca_dim}{PCA dimensions to plot as a vector of length 2. Default is \code{c(1,2)}}

\item{qc_types}{qc types to plot. Default is \code{c("SPL", "BQC", "TQC", "NIST", "LTR")}}

\item{label_k_mad}{Show analysis_id label for points outside k * mad of any of two defined PCA dimensions. Default is \code{3}. Set to \code{NULL} to supress labels.}

\item{log_transform}{log-transform data before calculating PCA. Default is \code{TRUE}}

\item{remove_istds}{Exlude ISTD features. Default is \code{TRUE}}

\item{min_median_signal}{Minimum median signal across all samples from all selected qc types. \code{NA} (default) will not filter any features based on signal.}

\item{point_size}{size of points. Default is \code{2}}

\item{point_alpha}{transparency of points}

\item{ellipse_alpha}{transparency of ellipse fill}

\item{font_base_size}{Base font size for plot text elements}

\item{hide_label_text}{Hide shared text (case-senstive) in labels. If it results in non-unique analysis_id's then an error will be raised.}
}
\value{
ggplot2 object
}
\description{
PCA plot for QC
}
