% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-responsecurves.R
\name{qc_plot_responsecurves}
\alias{qc_plot_responsecurves}
\title{Plot Response Curves}
\usage{
qc_plot_responsecurves(
  data = NULL,
  variable = "intensity",
  filter_data = FALSE,
  output_pdf = FALSE,
  include_feature_filter = "",
  exclude_feature_filter = "",
  max_regression_value = NA,
  path = "",
  rows_page = 4,
  cols_page = 5,
  specific_page = NA,
  point_size = 1.5,
  line_width = 0.7,
  scaling_factor = 1,
  page_orientation = "LANDSCAPE",
  font_base_size = 7,
  show_progress = TRUE,
  return_plots = FALSE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object containing the dataset with sample amounts and detector signals. Required.}

\item{output_pdf}{Logical. If \code{TRUE}, saves the generated plots as a PDF file. Defaults to \code{FALSE}.}

\item{include_feature_filter}{A regex pattern to filter and include features that match the criteria. If omitted, all features are considered.}

\item{exclude_feature_filter}{A regex pattern to filter and exclude features that match the criteria. If omitted, no features are excluded.}

\item{max_regression_value}{The maximum x value (analyzed amount) for which the regression line is fitted. Defaults to \code{NA}, considering all data points.}

\item{path}{The file path for saving the PDF. Defaults to an empty string, meaning no file is saved.}

\item{rows_page}{The number of rows of plots per page for the PDF output. Defaults to 4.}

\item{cols_page}{The number of columns of plots per page for the PDF output. Defaults to 5.}

\item{specific_page}{An integer specifying a particular page to plot. If \code{NA}, all pages are plotted.}

\item{point_size}{Numeric value specifying the size of points in the plot. Defaults to 1.5.}

\item{line_width}{Numeric value specifying the width of regression lines. Defaults to 0.7.}

\item{scaling_factor}{A universal scaling factor for fonts, symbols, and lines. Defaults to 1.}

\item{page_orientation}{The orientation of the PDF paper, either \code{"LANDSCAPE"} or \code{"PORTRAIT"}. Defaults to \code{"LANDSCAPE"}.}

\item{font_base_size}{The base font size for text in the plots. Defaults to 7.}

\item{return_plots}{Logical. If \code{TRUE}, returns the plots as a list of \code{ggplot2} objects. Each item represents a page of plots. Defaults to \code{FALSE}.}

\item{use_filtered_data}{Logical. Indicates whether to use quality control (QC) filtered data (\code{TRUE}) or the raw data (\code{FALSE}). Defaults to \code{FALSE}.}

\item{plot_variable}{The variable name to plot on the y-axis, usually a measure of intensity. Defaults to \code{"intensity"}.}

\item{display_progress}{Logical. Show a progress bar during plot creation if set to \code{TRUE}. Defaults to \code{TRUE}.}
}
\value{
If \code{return_plots} is \code{TRUE}, a list of \code{ggplot2} objects is returned. Otherwise, the function may return a plot output and/or save a PDF, depending on the \code{output_pdf} parameter.
}
\description{
Generates linear regression plots for each measured feature within a dataset. The function allows for customization of the plot's appearance and format, and it can filter features based on specified criteria.
}
