% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-runorder.R
\name{qc_plot_runboxplot}
\alias{qc_plot_runboxplot}
\title{Sample sequence Boxplots}
\usage{
qc_plot_runboxplot(
  data,
  relative_log_abundances,
  plot_var,
  use_qc_filtered_data,
  min_feature_intensity,
  qc_types = c("BQC|TQC|SPL|NIST|LTR"),
  ignore_outliers = TRUE,
  feature_filter = "",
  show_batches = TRUE,
  batches_as_shades = FALSE,
  batch_line_color = "red",
  batch_shading_color = "grey70",
  base_font_size = 8
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{relative_log_abundances}{Plot as RLA (Relative Log Abundnance Plot)}

\item{plot_var}{Variable to plot}

\item{use_qc_filtered_data}{Use QC-filtered data}

\item{min_feature_intensity}{Exclude features with median signal below this value}

\item{qc_types}{QC types to be plotted}

\item{ignore_outliers}{Exclude outlier values based on 3x IQR fences}

\item{feature_filter}{Filter text to select specific features (regex string)}

\item{show_batches}{Show batches}

\item{batches_as_shades}{Show batches as shades}

\item{batch_line_color}{batch separator color}

\item{batch_shading_color}{batch shade color}

\item{base_font_size}{base font size for plots (default is 8)}
}
\value{
ggplot object
}
\description{
Sample sequence Boxplots
}
