% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-runorder.R
\name{qc_plot_runsequence}
\alias{qc_plot_runsequence}
\title{RunSequence plot}
\usage{
qc_plot_runsequence(
  data,
  qc_types = NA,
  show_batches = TRUE,
  show_timestamp = FALSE,
  add_info_title = TRUE,
  single_row = FALSE,
  segment_thickness = 0.5,
  batches_as_shades = FALSE,
  batch_line_color = "#b6f0c5",
  batch_shading_color = "grey85",
  scale_factor = 1,
  show_outlier = TRUE,
  base_font_size = 8
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types to be shown, \code{NA} displays all available QC/Sample types. Can be a vector with QC types or a string with regex pattern.}

\item{show_batches}{Show batches}

\item{show_timestamp}{Use acquisition time stamp as x-axis}

\item{add_info_title}{Add title with experiment title and analysis date and times}

\item{single_row}{Show all qc types in a single row}

\item{segment_thickness}{Linewidth of the segments}

\item{batches_as_shades}{Show batches as shades}

\item{batch_line_color}{batch separator color}

\item{batch_shading_color}{batch shade color}

\item{scale_factor}{Overall plot scale factor}

\item{show_outlier}{Should samples defined as outlier as separate row}

\item{base_font_size}{base font size}
}
\value{
ggplot object
}
\description{
RunSequence plot
}
