% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-filtering.R
\name{qc_plot_summary}
\alias{qc_plot_summary}
\title{Plot summary of feature QC filtering}
\usage{
qc_plot_summary(
  data,
  use_batches = "summarise",
  with_venn_diag = TRUE,
  user_defined_keeper = FALSE,
  font_base_size = 8
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{use_batches}{How batches should be used, either across batches (ignoring batches), plot batch individually, or summarize batches (median)}

\item{with_venn_diag}{Include Venn diagram of features failing S/B, CVa, and linearity}

\item{user_defined_keeper}{Include user-defined feature inclusion list, even they did not pass QC filtering}

\item{font_base_size}{font size of plots}
}
\value{
ggplot2 object
}
\description{
Plot summary of feature QC filtering
}
