% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-filtering.R
\name{qc_plot_summary_classes}
\alias{qc_plot_summary_classes}
\title{Plot summary of feature QC filtering per class}
\usage{
qc_plot_summary_classes(
  data,
  include_qualifier = FALSE,
  include_istd = FALSE,
  user_defined_keeper = FALSE,
  font_base_size = 8
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{include_qualifier}{Include qualifier features. Is not used when \code{filter_data = TRUE} was applied.}

\item{include_istd}{Include internal standard features. Default is \code{TRUE}. Is not used when \code{filter_data = TRUE} was applied.}

\item{user_defined_keeper}{Include user-defined feature inclusion list, even they did not pass QC filtering}

\item{font_base_size}{font size of plots}
}
\value{
ggplot2 object
}
\description{
Plot summary of feature QC filtering per class
}
