% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-msdata.R
\name{rawdata_import_mrmkit}
\alias{rawdata_import_mrmkit}
\title{Reads a long CSV file with Feature Intensities}
\usage{
rawdata_import_mrmkit(
  file,
  analysis_id_col = NULL,
  feature_id_col = NULL,
  silent = FALSE
)
}
\arguments{
\item{file}{File name and path of a plain long-format CSV file}

\item{analysis_id_col}{Column to be used as analysis_id}

\item{feature_id_col}{Column to be used feature_id}

\item{silent}{Suppress messages}
}
\value{
A tibble in the long format
}
\description{
Reads a long CSV file with Feature Intensities
}
