% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{read_analysisresult_table}
\alias{read_analysisresult_table}
\title{Reads a generic wide .csv or .xls sheet with analysis results data}
\usage{
read_analysisresult_table(
  file,
  value_type = c("area", "height", "intensity", "norm_area", "norm_height",
    "norm_intensity", "feature_conc", "rt", "fwhm", "width"),
  sheet = "",
  silent = FALSE
)
}
\arguments{
\item{file}{File name and path of a plain wide-format CSV or XLS file}

\item{value_type}{Parameter type of the feature, i.e. "area", "height", "intensity", "norm_area", "norm_height", "norm_intensity", "feature_conc", "rt", "fwhm" or "width"}

\item{sheet}{Sheet name in case an Excel file (.xls, .xlsx, .xlsm) is imported}

\item{silent}{Suppress messages}
}
\value{
A tibble in the long format
}
\description{
Imports a tables with the column being the analysis identifier (e.g. sample name) and the following colums being measured features. The values in the table
}
