% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_analysisresults_long_format}
\alias{read_analysisresults_long_format}
\title{Reads a long CSV file with Feature Intensities}
\usage{
read_analysisresults_long_format(
  file,
  analysis_id_col = NULL,
  feature_name_col = NULL,
  silent = FALSE
)
}
\arguments{
\item{file}{File name and path of a plain long-format CSV file}

\item{silent}{Suppress messages}

\item{field}{Peak parameter (e.g. Area, RT)}
}
\value{
A tibble in the long format
}
\description{
Reads a long CSV file with Feature Intensities
}
