% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_long_table_csv}
\alias{read_long_table_csv}
\title{Reads a long CSV file with Feature Intensities}
\usage{
read_long_table_csv(file, field, silent = FALSE)
}
\arguments{
\item{file}{File name and path of a plain long-format CSV file}

\item{field}{Peak parameter (e.g. Area, RT)}

\item{silent}{Suppress messages}
}
\value{
A tibble in the long format
}
\description{
Reads a long CSV file with Feature Intensities
}
