% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-class.R
\name{read_masshunter_csv}
\alias{read_masshunter_csv}
\title{Reads an Agilent MassHunter Quant CSV file}
\usage{
read_masshunter_csv(data, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{filename}
}
\value{
MidarExperiment object
}
\description{
Imports a .csv file with \verb{Agilent MassHunter Quantitative Analysis} results.
Samples should be in rows, features/compounds in columns and must contain either peak areas, peak heights or intensities.
Additional columns, such as RT (rentention time), FWHM, PrecursorMZ, and CE will be imported and available from the \code{MidarExperiment} object for downstream analyses
}
\examples{
csvfile <- system.file("extdata", "Example_MHQuant_1.csv", package = "midar")
mexp <- MidarExperiment()
mexp <- read_masshunter_csv(mexp, csvfile)
mexp
}
