% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_MassHunterCSV}
\alias{read_MassHunterCSV}
\title{Read and convert an Agilent MassHunter Quant CSV result file}
\usage{
read_MassHunterCSV(filename, silent = FALSE)
}
\arguments{
\item{filename}{File path of MassHunter Quant CSV file}

\item{silent}{Suppress messages}
}
\value{
A tibble in the long format
}
\description{
Read and convert an Agilent MassHunter Quant CSV result file
}
