% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-metadata.R
\name{read_metadata_midarxlm}
\alias{read_metadata_midarxlm}
\title{Reads and parses metadata provided by the MSOrganizer EXCEL  template.}
\usage{
read_metadata_midarxlm(path, trim_ws = TRUE)
}
\arguments{
\item{path}{File path of the MSOrganizer EXCEL template (*.xlm)}

\item{trim_ws}{Trim all white spaces and double spaces}
}
\value{
A list with tibbles containing different metadata
}
\description{
Requires version 1.9.1 of the template
NOTES
\itemize{
\item if no sample_type is defined then SPL will be assigned
\item if valid_analysis is left blank for all analyses then samples then it will be replace by TRUE
\item if valid_analysis is undefined for one or more, but all all samples, then an error will be returned
}
}
