% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_mrmkit_result}
\alias{read_mrmkit_result}
\title{Reads a long CSV file with Feature Intensities}
\usage{
read_mrmkit_result(filename, use_normalized_data = FALSE, silent = FALSE)
}
\arguments{
\item{filename}{File name of the MRMkit result file (*.tsv or *.csv)}

\item{use_normalized_data}{Import raw peak areas or normalized peak areas from the file}

\item{silent}{No comments printed}
}
\value{
A tibble in the long format
}
\description{
Reads a long CSV file with Feature Intensities
}
