% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_MRMkit_results}
\alias{read_MRMkit_results}
\title{Reads a long CSV file with Feature Intensities}
\usage{
read_MRMkit_results(data, raw_area_file, final_results_file, silent = FALSE)
}
\arguments{
\item{data}{MidarExperiment object}

\item{raw_area_file}{file path of MRMkit csv output with raw peak areas}

\item{final_results_file}{file path of MRMkit csv output with final processed normalized peak areas}

\item{silent}{No comments printed}
}
\value{
A tibble in the long format
}
\description{
Reads a long CSV file with Feature Intensities
}
