% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_msorganizer_xlm}
\alias{read_msorganizer_xlm}
\title{Reads and parses metadata provided by the MSOrganizer EXCEL  template.}
\usage{
read_msorganizer_xlm(filename, trim_ws = TRUE)
}
\arguments{
\item{filename}{File path of the MSOrganizer EXCEL template (*.xlm)}

\item{trim_ws}{Trim all white spaces and double spaces}
}
\value{
A list of tibbles with different metadata
}
\description{
Requires version 1.9.1 of the template
}
