% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_table_wide}
\alias{read_table_wide}
\title{Reads a wide CSV or XLSX sheet with Feature Values}
\usage{
read_table_wide(data, file, field, sheet = "", silent = FALSE)
}
\arguments{
\item{data}{MidarExperiment()}

\item{file}{File name and path of a plain wide-format CSV or XLS file}

\item{field}{Peak parameter (e.g. Area, RT)}

\item{sheet}{Sheet name}

\item{silent}{Suppress messages}
}
\value{
A tibble in the long format
}
\description{
Reads a wide CSV or XLSX sheet with Feature Values
}
