% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{report_write_csv}
\alias{report_write_csv}
\title{Export any parameter to a wide-format table}
\usage{
report_write_csv(
  data,
  path,
  variable = c("area", "height", "intensity", "norm_intensity", "response", "conc",
    "conc_raw", "rt", "fwhm"),
  filter_data,
  qc_types = c("SPL", "BQC", "TQC", "NIST", "LTR", "PBLK", "SBLK", "UBLK", "MBLK"),
  include_qualifier = NA,
  include_istd = NA,
  add_qctype = NA
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{File name with path of exported CSV fil}

\item{variable}{Variable to be exported}

\item{filter_data}{Use QC-filtered data, based on criteria set via \code{qc_apply_feature_filter()}. Overwrites \code{include_qualifier} and \code{include_istd}.}

\item{qc_types}{QC type to plot. When qc_types us NA or NULL, all available QC types are plotted.}

\item{include_qualifier}{Include qualifier features. Is not used when \code{filter_data = TRUE} was applied.}

\item{include_istd}{Include internal standard features. Default is \code{TRUE}. Is not used when \code{filter_data = TRUE} was applied.}

\item{add_qctype}{Add the QC type as column}
}
\value{
A tibble with the exported data
}
\description{
Export any parameter to a wide-format table
}
