% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{report_write_qc_metrics}
\alias{report_write_qc_metrics}
\title{Save the QC table to a CSV file}
\usage{
report_write_qc_metrics(data = NULL, path)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{File name with path of exported CSV file}
}
\value{
A tibble with the exported dataset
}
\description{
Save the QC table to a CSV file
}
