% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{report_write_xlsx}
\alias{report_write_xlsx}
\title{Writes all a data processing report to an EXCEL file}
\usage{
report_write_xlsx(data, path)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{File name and path of the Excel file}
}
\description{
Writes all a data processing report to an EXCEL file
}
