% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{save_feature_qc_metrics}
\alias{save_feature_qc_metrics}
\title{Save Feature QC Metrics to CSV}
\usage{
save_feature_qc_metrics(data = NULL, path)
}
\arguments{
\item{data}{A MidarExperiment object containing the QC metrics.}

\item{path}{A string specifying the file path where the CSV file will be saved.}
}
\value{
A tibble with the QC metrics that have been exported.
}
\description{
This function exports the feature information and QC (Quality Control) metrics
from a MidarExperiment object to a CSV file.
}
