% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{save_metadata_templates}
\alias{save_metadata_templates}
\title{Saves a Excel (xlsx) file with metadata templates}
\usage{
save_metadata_templates(path = "metadata_template.xlsx")
}
\arguments{
\item{path}{File path where the XLSX file with templates will be saved.
If left empty (default), the file will be saved in the current working directory
under the file "metadata_template.xlsx"}
}
\description{
This function saves a XLSX file with metadata template to the specified location.
}
