% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{save_report_xlsx}
\alias{save_report_xlsx}
\title{Writes all a data processing report to an EXCEL file}
\usage{
save_report_xlsx(data = NULL, path)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{File name and path of the Excel file}
}
\description{
Writes all a data processing report to an EXCEL file
}
