% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-managment.R
\name{set_intensity_var}
\alias{set_intensity_var}
\title{Set default variable to be used as feature raw signal value}
\usage{
set_intensity_var(
  data = NULL,
  variable_name,
  auto_select = FALSE,
  warnings = TRUE,
  ...
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable_name}{Feature variable to be used as default feature intensity for downstream processing.}

\item{auto_select}{If \code{TRUE} then the first available of these will be used as default: "intensity", "response", "area", "height".}

\item{warnings}{Suppress warnings}

\item{...}{Feature variables to best search for one-by-one when \code{auto-detect = TRUE}}
}
\value{
MidarExperiment object
}
\description{
Sets the raw signal variable used for calculations starting from raw signal
values (i.e., normalization) Note that this set variable must be part of the
orginally imported data. Processed data variables (e.g., normalized intensities
and concentrations) can not be set as default feature intensity variable.
}
