% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-EDA.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Get FC and p values from a t test contrast}
\usage{
volcano_plot(
  d_FC,
  p_adjust,
  sig_FC_min,
  sig_p_value_min,
  symmetric_x,
  x_min,
  x_max,
  point_size,
  point_transparency,
  scale_factor,
  scale_factor_species_label,
  squared = TRUE,
  axistick = 1
)
}
\arguments{
\item{d_FC}{data.frame with FEATURE_NAME, p and log2FC}

\item{p_adjust}{calculate and use FDR}

\item{sig_FC_min}{|FC| treshold}

\item{sig_p_value_min}{P value treshold}

\item{symmetric_x}{make x axis symetric around 0}

\item{x_min}{x min}

\item{x_max}{x max}

\item{point_size}{point size}

\item{point_transparency}{point alpha}

\item{scale_factor}{scale factor for fonts}

\item{scale_factor_species_label}{scale factor point labels}

\item{squared}{make plot a square}

\item{axistick}{axis ticks}
}
\value{
tibble with p values, fdr and log2FC
}
\description{
Get FC and p values from a t test contrast
}
