% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-export.R
\name{writeReportXLS}
\alias{writeReportXLS}
\title{Writes all a data processing report to an EXCEL file}
\usage{
writeReportXLS(data, path)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{File name and path of the Excel file}
}
\description{
Writes all a data processing report to an EXCEL file
}
