% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-outlier.R
\name{analysis_outlier_detection}
\alias{analysis_outlier_detection}
\title{Get list of analyses classified as technical outliers}
\usage{
analysis_outlier_detection(
  data = NULL,
  variable = c("feature_intensity", "feature_norm_intensity", "feature_conc"),
  qc_types = c("BQC", "TQC", "SPL"),
  pca_component,
  fence_multiplicator,
  summarize_fun = c("pca", "rma"),
  outlier_detection = c("sd", "mad"),
  log_transform = TRUE,
  print_outliers = TRUE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable}{Feature variable used for outlier detection}

\item{qc_types}{QC types included in the outlier detection}

\item{pca_component}{PCA component to be used}

\item{fence_multiplicator}{Multiplicator for SD or MAD, respectively.}

\item{summarize_fun}{Function used to summarize the features, either "pca" based on PCA, or "rma" based on mean relative abundance (RMA) of all features}

\item{outlier_detection}{Outlier detection method, either based on SD ("sd") or MAD ("mad")}

\item{log_transform}{Log-transform data for outlier detection}

\item{print_outliers}{Print analysis_id of outliers to the console}
}
\value{
MidarExperiment object
}
\description{
Retrieves IDs os analysis technical outliers, i.e, analyses with systematic differences in multiple features.
This is based on SD or MAD fences from the mean or median, respectively, from the principal components of a PCA, or of the relative mean abundance (RMA)
}
