% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{apply_qc_filter}
\alias{apply_qc_filter}
\title{Filter dataset according to QC and other criteria}
\usage{
apply_qc_filter(
  data,
  exclude_technical_outliers = FALSE,
  min_intensity_bqc = NA,
  min_intensity_tqc = NA,
  min_intensity_spl = NA,
  min_signal_blank_ratio = NA,
  min_cv_conc_bqc = NA,
  min_cv_conc_tqc = NA,
  max_dratio_conc_bqc = NA,
  max_dratio_conc_tqc = NA,
  min_response_rsquare = NA,
  rqc_curve_used_for_filt = NA,
  keep_quantifier_only = TRUE,
  exclude_istds = TRUE,
  features_to_keep = NULL
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{exclude_technical_outliers}{Remove samples classified as outliers}

\item{min_intensity_bqc}{Minimum median signal intensity of BQC}

\item{min_intensity_tqc}{Minimum median signal intensity of TQC}

\item{min_intensity_spl}{Minimum median signal intensity of study samples (SPL)}

\item{min_signal_blank_ratio}{= Signal-to-Blank ratio. Calculated from the median of study samples and the median of the Process Blank (PBLK)}

\item{min_cv_conc_bqc}{= Maximum \%CV of BQC}

\item{min_cv_conc_tqc}{Maximum \%CV of TQC}

\item{max_dratio_conc_bqc}{D-ratio defined as CV_BQC/CV_SPL}

\item{max_dratio_conc_tqc}{D-ratio defined as CV_TQC/CV_SPL}

\item{min_response_rsquare}{= Minimum r squared of RQC curve defined under \code{rqc_curve_used_for_filt}}

\item{rqc_curve_used_for_filt}{Name of RQC curve as string, or index number of curve to use for filtering (first curve is 1)}

\item{keep_quantifier_only}{Remove features where Quantifier is set to FALSE.}

\item{exclude_istds}{Remove Internal Standards (ISTD). If set to FALSE, meaning ISTDs will be included, then \code{min_signal_blank_ratio} is ignored, because the the S/B is based on Processed Blanks (PBLK) that contain ISTDs.}

\item{features_to_keep}{Features that must be kept, even if they did not meet the given QC criteria}
}
\value{
MidarExperiment object
}
\description{
Filter dataset according to QC parameter criteria, remove features that are internal standards (ISTDs) or not annotated as quantifier (optional).
Exclude features and analyses that were annotated as not valid in the metadata (valid_integration, valid_analysis).

Note: When \code{exclude_istds} is FALSE, then \code{min_signal_blank_ratio} is ignored, because the the S/B is based on Processed Blanks (PBLK) that contain ISTDs.
}
