% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-metadata.R
\name{assert_metadata}
\alias{assert_metadata}
\title{Add metadata an MidarExperiment object}
\usage{
assert_metadata(data, metadata, excl_unannotated_analyses, ignore_warnings)
}
\arguments{
\item{data}{MidarExperiment object}

\item{excl_unannotated_analyses}{Exclude analyses (samples) that have no matching metadata}

\item{path}{List of tibbles or data.frames containing analysis, feature, istd, response curve tables}

\item{analysis_sequence}{Must by any of: "timestamp", "resultfile" or "metadata". Defines how the analysis order is determined. Default is "timestamp", when not available the sequence in the analysis results are used.}
}
\value{
metadata list
}
\description{
Metadata provided as a list of tibbles will validates for consistency again loaded analysis data of the provided MidarExperiment object and then transfered.
}
