% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-calibrations.R
\name{calc_calibration_results}
\alias{calc_calibration_results}
\title{Calculate external calibration curve results}
\usage{
calc_calibration_results(
  data = NULL,
  overwrite_metadata_fit = FALSE,
  fit_method = c("linear", "quadratic"),
  fit_weighting = c(NA, "none", "1/x", "1/x^2"),
  missing_error = TRUE,
  ignore_istd = FALSE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object}

\item{overwrite_metadata_fit}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the function will ignore any fit method and weighting settings defined in the metadata and use the \code{fit_method} and \code{fit_weighting} values for all analytes.}

\item{fit_method}{A character string indicating the default regression fit method to use for the calibration curve. Must be one of \code{"linear"} or \code{"quadratic"}. This method will be used if no specific fit method is defined for a feature in the metadata.}

\item{fit_weighting}{A character string indicating the default weighting method for the regression points in the calibration curve. Must be one of \code{"none"}, \code{"1/x"}, or \code{"1/x^2"}. If no specific weighting method is defined for a feature in the metadata, this method will be used.}

\item{missing_error}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, an error will be raised if one or more ISTD concentrations and sample/ISTD amounts are missing. Default is \code{TRUE}.}

\item{ignore_istd}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE}, ISTD values with missing concentrations that are not used in any feature quantification will be ignored. Default is \code{FALSE}.}
}
\value{
A modified \code{MidarExperiment} object with updated concentration values for the analytes, based on the calibration curve calculations.
}
\description{
Concentrations are determined using ISTD-normalized intensities and external calibration curves. The determined concentration unit is based on values defined in the column \code{concentration_unit} of the qc_concentration metadata.
}
\details{
The fit model (e.g., linear or quadratic) and the weighting method (e.g., none, 1/x, 1/x^2) used for the calibration curve can either be defined for each analyte in the metadata, or, if missing, the default values provided as arguments will be used.
}
