% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{$,MidarExperiment-method}
\alias{$,MidarExperiment-method}
\title{Access Slots of a MidarExperiment Object via $ Syntax}
\usage{
\S4method{$}{MidarExperiment}(x, name)
}
\arguments{
\item{x}{MidarExperiment object}

\item{name}{MidarExperiment slot}
}
\value{
Value with a variable or a tibble
}
\description{
$ syntax can be used to as a shortcut for getting specific variables and results from a MidarExperiment object
}
\examples{
mexp <- MidarExperiment(title = "Test Experiment", analysis_type = "lipidomics")
mexp$analysis_type
mexp$title
mexp$annot_analyses
}
