% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorr.R
\name{corr_batch_centering}
\alias{corr_batch_centering}
\title{Batch centering}
\usage{
corr_batch_centering(
  data,
  qc_types,
  use_raw_concentrations = FALSE,
  center_fun = "median"
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types used for batch correction}

\item{use_raw_concentrations}{Apply to unadjusted (raw) concentration. Default is FALSE, which means previously drift-corrected concentrations will be used if available, otherwise unadjusted concentrations will be used}

\item{center_fun}{Function used to center. Default is "median".}
}
\value{
MidarExperiment object
}
\description{
A short description...
#' @details
Additional details...
}
