% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_drift-batch.R
\name{corr_batcheffects}
\alias{corr_batcheffects}
\title{Batch centering}
\usage{
corr_batcheffects(
  data,
  qc_types,
  correct_location = TRUE,
  correct_scale = FALSE,
  overwrite_batch_corr = TRUE,
  log_transform = TRUE,
  ...
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types used for batch correction}

\item{correct_location}{Align locations (median) of batches}

\item{correct_scale}{Scale batches to the same level}

\item{overwrite_batch_corr}{Overwrite previous batch correction or apply on top existing batch correction}

\item{log_transform}{Log transform the data internally for correction. Will not transform the final results.}

\item{...}{Other parameters for batch correction function. Currently not in use.}
}
\value{
MidarExperiment object
}
\description{
A short description...
#' @details
Additional details...
}
