% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-drift-batch.R
\name{corr_drift_fun}
\alias{corr_drift_fun}
\title{Drift Correction by Custom Function}
\usage{
corr_drift_fun(
  data,
  smooth_fun,
  qc_types,
  log2_transform = TRUE,
  within_batch,
  apply_conditionally,
  apply_conditionally_per_batch = TRUE,
  max_cv_ratio_before_after = 1,
  use_uncorrected_if_fail = TRUE,
  ignore_istd = TRUE,
  feature_list = NULL,
  ...
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{smooth_fun}{Function that performs drift correction. Function need to have following parameter \code{data} (\code{MidarExperiment}), \code{qc_typeS} (one or more strings), and \code{span_width} (numerical).
Function needs to return a numerical vector with the length of number of rows in \code{data}. In case functions fails a vector with NA_real_ needs be returned}

\item{qc_types}{QC types used for drift correction}

\item{log2_transform}{Log transform the data for correction when \code{TRUE} (the default). Note: log transformation is solely applied internally for smoothing, results will not be be log-transformed. Log transformation may result in more robust smoothing that is less sensitive to outlier.}

\item{within_batch}{Apply to each batch separately if \code{TRUE} (the default)}

\item{apply_conditionally}{Apply drift correction to all species if \code{TRUE}, or only when sample CV after smoothing changes below a threshold defined via \code{max_cv_ratio_before_after}}

\item{apply_conditionally_per_batch}{When \code{apply_conditionally = TRUE}, correction is conditionally applied per batch when \code{TRUE} and across all batches when \code{FALSE}}

\item{max_cv_ratio_before_after}{Only used when \code{apply_conditionally = TRUE}. Maximum allowed ratio of sample CV change before and after smoothing for the correction to be applied.
A value of 1 (the default) indicates the CV needs to improve or remain unchanged after smoothing so that the conditional smoothing is applied. A value of < 1 means that CV needs to improve, a value of e.g. 1.20 that the CV need to improve or get worse by max 1.20-fold after smoothing.}

\item{use_uncorrected_if_fail}{In case the smoothing function fails for a species, then use original (uncorrected) data when \code{TRUE} (the default) or return \code{NA} for all analyses of the feature where the fit failed.}

\item{ignore_istd}{Do not apply corrections to ISTDs}

\item{feature_list}{Subset the features for correction whose names matches the specified text using regular expression. Default is \code{NULL} which means all features are selected.}

\item{...}{Parameters specific for the smoothing function}
}
\value{
MidarExperiment object
}
\description{
Function to correct for run-order drifts within or across batches via a provided custom function
#' @details
The drift correction function needs to be provided by the user. See \code{smooth_fun} for details.
}
