% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-drift-batch.R
\name{corr_drift_fun}
\alias{corr_drift_fun}
\title{Drift Correction by Custom Function}
\usage{
corr_drift_fun(
  data = NULL,
  smooth_fun,
  variable,
  reference_qc_types,
  within_batch,
  replace_previous = TRUE,
  log_transform_internal = TRUE,
  conditional_correction = FALSE,
  cv_ratio_threshold = 1,
  use_original_if_fail = TRUE,
  ignore_istd = TRUE,
  feature_list = NULL,
  recalc_trend_after = FALSE,
  show_progress = TRUE,
  ...
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{smooth_fun}{Function that performs drift correction. Function need to have following parameter \code{data} (\code{MidarExperiment}), \code{reference_qc_types} (one or more strings), and \code{span_width} (numerical).
Function needs to return a numerical vector with the length of number of rows in \code{data}. In case functions fails a vector with NA_real_ needs be returned}

\item{variable}{The variable to be corrected for drift effects. Must be one of "intensity", "norm_intensity", or "conc"}

\item{reference_qc_types}{QC types used for drift correction}

\item{within_batch}{Apply to each batch separately if \code{TRUE} (the default)}

\item{replace_previous}{Logical. Replace previous correction (\code{TRUE}), or adds on top of previous correction (\code{FALSE}). Default is \code{TRUE}.}

\item{log_transform_internal}{Apply log transformation internally for smoothing if \code{TRUE} (default). This enhances robustness against outliers but does not affect the final data, which remains untransformed.}

\item{conditional_correction}{Determines whether drift correction should be applied to all species unconditionally (\code{TRUE}) or only when the sample CV after smoothing is below the ratio threshold specified by \code{cv_ratio_threshold}.}

\item{cv_ratio_threshold}{This parameter defines the maximum allowable change (ratio) in the coefficient of variation (CV) ratio of samples before and after smoothing for the correction to be applied.
A value of 1 (the default) requires the CV to improve, while a value above 1 allows the CV to improve, or to degrade by a maximum of the indicated fold-difference times post-smoothing.}

\item{use_original_if_fail}{Determines the action when smoothing fails for a feature. If TRUE (default), the original data is used; if FALSE, the result for each analysis is NA.}

\item{ignore_istd}{Do not apply corrections to ISTDs}

\item{feature_list}{Subset the features for correction whose names matches the specified text using regular expression. Default is \code{NULL} which means all features are selected.}

\item{recalc_trend_after}{Recalculate trend post-drift correction for \code{plot_qc_runscatter()}. This will double calculation time.}

\item{show_progress}{Show progress bar. Default = `TRUE.}

\item{...}{Parameters specific for the smoothing function}
}
\value{
MidarExperiment object
}
\description{
Function to correct for run-order drifts within or across batches via a provided custom function
#' @details
The drift correction function needs to be provided by the user. See \code{smooth_fun} for details.
}
