% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorr.R
\name{corr_drift_gaussiankernel}
\alias{corr_drift_gaussiankernel}
\title{Drift Correction by Gaussian Kernel Smoothing}
\usage{
corr_drift_gaussiankernel(
  data,
  qc_types,
  bandwidth,
  log2_transform = TRUE,
  within_batch,
  apply_conditionally,
  apply_conditionally_per_batch = TRUE,
  feature_list = NULL,
  min_sample_cv_ratio_before_after = 1,
  use_uncorrected_if_fit_failed = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types used for drift correction. Typically includes the study samples (\code{SPL}).}

\item{bandwidth}{Kernel bandwidth}

\item{log2_transform}{log2 transform data during correction (Default is TRUE). Log transformation is only used for the fit, results will not be transformed.}

\item{within_batch}{Correct each batch separately (Default is TRUE)}

\item{apply_conditionally}{Apply drift correction to all species or conditionally based on 'min_sample_cv_ratio_before_after'}

\item{apply_conditionally_per_batch}{Apply correction conditionally using min_sample_cv_ratio_before_after criteriaper batch or across batches}

\item{feature_list}{Apply correction only to species matching (RegEx)}

\item{min_sample_cv_ratio_before_after}{Maximum sample CV change for correction to be applied}

\item{use_uncorrected_if_fit_failed}{If fit function fails, use orginal (uncorrected) data or return NA for all analyses of this feature}
}
\value{
MidarExperiment object
}
\description{
Function to correct for run-order drifts within or across batches using gaussian kernel smoothing. This is typically used to smooth based on the study samples.
#' @details
}
