% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorr.R
\name{corr_drift_loess}
\alias{corr_drift_loess}
\title{Drift Correction by LOESS Smoothing}
\usage{
corr_drift_loess(
  data,
  qc_types,
  within_batch,
  apply_conditionally,
  apply_conditionally_per_batch = TRUE,
  log2_transform = TRUE,
  span = 0.75,
  feature_list = NULL,
  min_sample_cv_ratio_before_after = 1,
  extrapolate = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types used for drift correction}

\item{within_batch}{Correct each batch separately (Default is TRUE)}

\item{apply_conditionally}{Apply drift correction to all species or conditionally based on 'min_sample_cv_ratio_before_after'}

\item{apply_conditionally_per_batch}{Apply correction conditionally using min_sample_cv_ratio_before_after criteriaper batch or across batches}

\item{log2_transform}{log2 transform data during correction (Default is TRUE). Log transformation is only used for the fit, results will not be transformed.}

\item{span}{Loess span width (default is 0.75)}

\item{feature_list}{Apply correction only to species matching (RegEx).}

\item{min_sample_cv_ratio_before_after}{Maximum sample CV change for correction to be applied}

\item{extrapolate}{Extrapolate loess smoothing. WARNING: Use with casre, it is generally not recommended to extrapolate outside of the range spanned by the QCs used for smoothing. See details below.}
}
\value{
MidarExperiment object
}
\description{
Function to correct for run-order drifts within or across batches using loess smoothing
}
\details{
Note that using extrapolation for loess smoothing is generally not recommended. Use this only if you must include samples or QCs that are outside of the range spanned by the QCs used for smoothing.
Cases where this may be necessary are when specific drifts occur in analysis sequence segments that are not spanned by QC, e.g. when the instrument broke down or suddenly changed its sensitivity.
Only use for samples  adjacent to the first or last QC and consult runscatter plots.
}
