% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-drift-batch.R
\name{correct_batch_centering}
\alias{correct_batch_centering}
\title{Batch centering}
\usage{
correct_batch_centering(
  data = NULL,
  qc_types,
  correct_location = TRUE,
  correct_scale = FALSE,
  overwrite = TRUE,
  calc_log_transform = TRUE,
  ...
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types used for batch correction}

\item{correct_location}{Align locations (median) of batches}

\item{correct_scale}{Scale batches to the same level}

\item{overwrite}{Overwrite previous batch correction or apply on top existing batch correction}

\item{calc_log_transform}{Log transform the data internally for correction. Note: this will not transform the final results.}

\item{...}{Other parameters for batch correction function. Currently not in use.}
}
\value{
MidarExperiment object
}
\description{
A short description...
#' @details
Additional details...
}
