% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-drift-batch.R
\name{correct_drift_gaussiankernel}
\alias{correct_drift_gaussiankernel}
\title{Drift Correction by Gaussian Kernel Smoothing}
\usage{
correct_drift_gaussiankernel(
  data = NULL,
  variable,
  reference_qc_types,
  within_batch = TRUE,
  replace_previous = TRUE,
  kernel_size,
  outlier_filter = FALSE,
  outlier_ksd = 5,
  location_smooth = TRUE,
  scale_smooth = FALSE,
  log_transform_internal = TRUE,
  recalc_trend_after = TRUE,
  conditional_correction = FALSE,
  feature_list = NULL,
  ignore_istd = TRUE,
  cv_ratio_threshold = 1,
  use_original_if_fail = FALSE,
  show_progress = TRUE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable}{The variable to be corrected for drift effects. Must be one of "intensity", "norm_intensity", or "conc"}

\item{reference_qc_types}{QC types used for drift correction. Typically includes the study samples (\code{SPL}).}

\item{within_batch}{Apply to each batch separately if \code{TRUE} (the default), or across all batches if \code{FALSE}.}

\item{replace_previous}{Logical. Replace previous correction (\code{TRUE}), or adds on top of previous correction (\code{FALSE}). Default is \code{TRUE}.}

\item{kernel_size}{Kernel bandwidth}

\item{outlier_filter}{Kernel Outlier filter}

\item{outlier_ksd}{Kernel K times standard deviation of data distribution}

\item{location_smooth}{Location parameter smoothing}

\item{scale_smooth}{Scale parameter smoothing}

\item{log_transform_internal}{Apply log transformation internally for smoothing if \code{TRUE} (default). This enhances robustness against outliers but does not affect the final data, which remains untransformed.}

\item{recalc_trend_after}{Recalculate trends after smoothing, used for plotting (e.g., in \code{plot_runscatter()})}

\item{conditional_correction}{Apply drift correction to all species if \code{TRUE}, or only when sample CV after smoothing changes below a threshold defined via \code{cv_ratio_threshold}}

\item{feature_list}{Subset the features for correction whose names matches the specified text using regular expression. Default is \code{NULL} which means all features are selected.}

\item{ignore_istd}{Do not apply corrections to ISTDs}

\item{cv_ratio_threshold}{Only used when \code{conditional_correction = TRUE}. Maximum allowed ratio of sample CV change before and after smoothing for the correction to be applied.
A value of 1 (the default) indicates the CV needs to improve or remain unchanged after smoothing so that the conditional smoothing is applied. A value of < 1 means that CV needs to improve, a value of e.g. 1.20 that the CV need to improve or get worse by max 1.20-fold after smoothing.}

\item{use_original_if_fail}{Determines the action when smoothing fails for a feature. If TRUE (default), the original data is used; if FALSE, the result for each analysis is NA.}

\item{show_progress}{Show progress bars. Set this to \code{FALSE} when rendering the notebook.}
}
\value{
MidarExperiment object
}
\description{
Function to correct for run-order drifts within or across batches using gaussian kernel smoothing (see \emph{Tan et al. (2020)}).
This is typically used to smooth based on the study samples. To avoid local biases and artefacts, this function should only be applied to analyses wit sufficient number of samples that were well randomized.
}
\references{
Teo G., Chew WS, Burla B, Herr D, Tai ES, Wenk MR, Torta F, & Choi H (2020). MRMkit: Automated Data Processing for Large-Scale Targeted Metabolomics Analysis. \emph{Analytical Chemistry}, 92(20), 13677–13682. \url{https://doi.org/10.1021/acs.analchem.0c03060}
}
