% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-drift-batch.R
\name{correct_drift_gaussiankernel}
\alias{correct_drift_gaussiankernel}
\title{Drift Correction by Gaussian Kernel Smoothing}
\usage{
correct_drift_gaussiankernel(
  data,
  qc_types,
  batch_wise = TRUE,
  kernel_size,
  outlier_filter = FALSE,
  outlier_ksd = 5,
  location_smooth = TRUE,
  scale_smooth = FALSE,
  calc_log_transform = TRUE,
  recalc_trend_after = TRUE,
  apply_conditionally = FALSE,
  apply_conditionally_per_batch = FALSE,
  feature_list = NULL,
  ignore_istd = TRUE,
  max_cv_ratio_before_after = 1,
  use_uncorrected_if_fail = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_types}{QC types used for drift correction. Typically includes the study samples (\code{SPL}).}

\item{batch_wise}{Apply to each batch separately if \code{TRUE} (the default)}

\item{kernel_size}{Kernel bandwidth}

\item{outlier_filter}{Kernel Outlier filter}

\item{outlier_ksd}{Kernel K times standard deviation of data distribution}

\item{location_smooth}{Location parameter smoothing}

\item{scale_smooth}{Scale parameter smoothing}

\item{calc_log_transform}{Log transform the data for correction when \code{TRUE} (the default). Note: log transformation is solely applied internally for smoothing, results will not be be log-transformed. Log transformation may result in more robust smoothing that is less sensitive to outlier.}

\item{apply_conditionally}{Apply drift correction to all species if \code{TRUE}, or only when sample CV after smoothing changes below a threshold defined via \code{max_cv_ratio_before_after}}

\item{apply_conditionally_per_batch}{When \code{apply_conditionally = TRUE}, correction is conditionally applied per batch when \code{TRUE} and across all batches when \code{FALSE}}

\item{feature_list}{Subset the features for correction whose names matches the specified text using regular expression. Default is \code{NULL} which means all features are selected.}

\item{ignore_istd}{Do not apply corrections to ISTDs}

\item{max_cv_ratio_before_after}{Only used when \code{apply_conditionally = TRUE}. Maximum allowed ratio of sample CV change before and after smoothing for the correction to be applied.
A value of 1 (the default) indicates the CV needs to improve or remain unchanged after smoothing so that the conditional smoothing is applied. A value of < 1 means that CV needs to improve, a value of e.g. 1.20 that the CV need to improve or get worse by max 1.20-fold after smoothing.}

\item{use_uncorrected_if_fail}{In case the smoothing function fails for a species, then use original (uncorrected) data when \code{TRUE} (the default) or return \code{NA} for all analyses of the feature where the fit failed.}
}
\value{
MidarExperiment object
}
\description{
Function to correct for run-order drifts within or across batches using gaussian kernel smoothing (see \emph{Tan et al. (2020)}).
This is typically used to smooth based on the study samples. To avoid local biases and artefacts, this function should only be applied to analyses wit sufficient number of samples that were well randomized.
}
\references{
Teo G., Chew WS, Burla B, Herr D, Tai ES, Wenk MR, Torta F, & Choi H (2020). MRMkit: Automated Data Processing for Large-Scale Targeted Metabolomics Analysis. \emph{Analytical Chemistry}, 92(20), 13677–13682. \url{https://doi.org/10.1021/acs.analchem.0c03060}
}
