% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isotopecorrections.R
\name{correct_interferences}
\alias{correct_interferences}
\title{Substract interferences contributed by another feature}
\usage{
correct_interferences(data)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable}{Name of Variable to be corrected. Default: \code{feature_intensity}.}
}
\value{
MidarExperiment object
}
\description{
The interference (e.g. isotope overlap or in-source fragments) is substracted using following formula:
\deqn{Value_corrected = Value_Feature - relative_contribution * Value_InterferingFeature}
}
