% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{data_import_mrmkit}
\alias{data_import_mrmkit}
\title{Imports MRMkit peak integration results}
\usage{
data_import_mrmkit(data, path, use_metadata, silent = FALSE)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{One or more file names with path, or a folder path, which case all *.csv files in this folder will be read.}

\item{use_metadata}{Import additional metadata columns (e.g. batch ID, sample type) and add to the \code{MidarExperiment} object}

\item{silent}{Suppress most notifications}
}
\value{
MidarExperiment object
}
\description{
Imports .tsv file(s) generated from \code{MRMkit} containing peak integration results.
The table format must be in the long format with columns for the raw data file name, feature ID, and the peak intensity and other parameters.
Additional information, such as retention time, FWHM, precursor/product MZ, and CE will also be imported and will available from the \code{MidarExperiment} object for downstream analyses.

When a path to a folder is provided, all .tsv files in that folder will be imported and merged into one raw dataset. This is useful, e.g. when importing datasets that are pre-processing in blocks resulting in different files.
Each feature and raw data file pair must only occur once within and across all .csv source data files, duplicated return an error.
}
\examples{
file_path <- system.file("extdata", "sPerfect_MRMkit.tsv", package = "midar")
mexp <- MidarExperiment()
mexp <- data_import_mrmkit(data = mexp, path = file_path, use_metadata = TRUE)
mexp
}
