% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-summarize.R
\name{data_sum_features}
\alias{data_sum_features}
\title{Sum up feature intensities per analyte}
\usage{
data_sum_features(data, qualifier_action = "include", feature_classes = NA)
}
\arguments{
\item{data}{MidarExperiment object}

\item{qualifier_action}{Character. How to handle qualifier features. To sum them up separately select "separate",}

\item{feature_classes}{Select feature classes to sum up. Default (NA) is to sum up all feature classes according to their analyte_id.
to include them in the sum if quantifier select "include", to not sum them up select "exclude".}
}
\value{
MidarExperiment object
}
\description{
This function sums up feature intensities per analyte_id.

THis is is useful when you have multiple features (e.g. adducts, isotopes, in-source fragments) or isomers that you
want to combine into a single analyte intensity value, such as LPC sn1 and sn2 species.

NOTE: This is still an experimental function! It will overwrite the feature_id in the dataset and analysis metadata
of featured that share same analyte_id. Currently the original feature_id is not backed up anywhere. Use with caution
and check results carefully!
}
