% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-drift-batch.R
\name{fun_gauss.kernel.smooth}
\alias{fun_gauss.kernel.smooth}
\title{Gaussian Kernel smoothing helper function}
\usage{
fun_gauss.kernel.smooth(tbl, qc_types, ...)
}
\arguments{
\item{tbl}{Table (\code{tibble} or \code{data.frame}) containing the fields \code{qc_type}, \code{x} (run order number), and \code{y} (variable)}

\item{qc_types}{QC types used for the smoothing (fit) by loess}

\item{...}{Additional parameters}
}
\value{
List with a \code{data.frame} containing original \code{x} and the smoothed \code{y} values, and a \code{boolean} value indicting whether the fit failed or not not.
}
\description{
Function for Gaussian kernel-based smoothing, for use by \code{corr_drift_fun}.
}
\author{
Hyung Won Choi
}
