% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchcorr.R
\name{fun_loess}
\alias{fun_loess}
\title{Loess smoothing helper function}
\usage{
fun_loess(tbl, qc_types, span_width, ...)
}
\arguments{
\item{tbl}{Table (\code{tibble} or \code{data.frame}) containing the fields \code{QC_TYPE}, \code{x} (run order number), and \code{y} (variable)}

\item{qc_types}{QC types used for the smoothing (fit) by loess}

\item{span_width}{Loess span width}

\item{...}{Additional parameters forwarded to Loess}
}
\value{
List with a \code{data.frame} containing original x and the smoothed y values, and a \code{boolean} value indicting whether the fit failed or not not.
}
\description{
Function for loess-based smoothing, for use by \code{corr_drift_fun}
}
