% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-managment.R
\name{get_analyticaldata}
\alias{get_analyticaldata}
\title{Get the annotated or the originally imported analytical data}
\usage{
get_analyticaldata(data = NULL, original = FALSE)
}
\arguments{
\item{data}{MidarExperiment object}

\item{original}{Boolean indicating whether to return the original imported data (\code{TRUE}) or the annotated data (\code{FALSE})}
}
\value{
A tibble with the analytical data in the long format
}
\description{
Get the annotated or the originally imported analytical data
}
