% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions-math.R
\name{get_iqr_tails}
\alias{get_iqr_tails}
\title{Get Tukey's IQR fences}
\usage{
get_iqr_tails(x, k = 1.5, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{k}{A numeric multiplier for the IQR. Default is \code{1.5}.}

\item{na.rm}{Logical. Should missing values be removed? Default is \code{FALSE}.}
}
\value{
A numeric vector of length 2: \code{c(lower_tail, upper_tail)}.
}
\description{
Computes the lower and upper boundaries based on Tukey's Interquartile Range (IQR)
fences from a numeric vector.
}
