% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-calibrations.R
\name{get_qc_bias_variability}
\alias{get_qc_bias_variability}
\title{Retrieve Calibration Regression Results}
\usage{
get_qc_bias_variability(
  data,
  qc_types = NA,
  sample_ids = NA,
  wide_format = "none",
  include_qualifier = FALSE,
  with_conc = TRUE,
  with_conc_target = TRUE,
  with_bias = TRUE,
  with_bias_abs = FALSE,
  with_conc_ratio = FALSE,
  with_cv_intra = TRUE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object containing the dataset and necessary annotations for calibration analysis.}

\item{qc_types}{A character vector specifying the QC types to include in the results, in addition to \code{CAL}. If not specified, all applicable QC types are included by default.}

\item{sample_ids}{A character vector specifying the sample IDs to include in the results. If not specified, all analyses regardless of their sample IDs are included by default.}

\item{wide_format}{Format of the output table. Must be one of \code{"none"}, \code{"features"}, or \code{"samples"}.
If \code{"none"}, the output is in long format. If \code{"features"}, the output is in wide format with features as columns.
If \code{"samples"}, the output is in wide format with samples as columns.}

\item{include_qualifier}{Logical. If \code{TRUE}, includes qualifier features in the results. Defaults to \code{FALSE}.}

\item{with_conc}{Logical. If \code{TRUE}, includes target and measured mean concentrations in the results. Defaults to \code{TRUE}.}

\item{with_conc_target}{Logical. If \code{TRUE}, includes target (know) concentration of the QC sample in the results. Defaults to \code{TRUE}.}

\item{with_bias}{Logical. If \code{TRUE}, includes percentage bias in the results. Defaults to \code{TRUE}.}

\item{with_bias_abs}{Logical. If \code{TRUE}, includes absolute bias in concentration units in the results. Defaults to \code{FALSE}.}

\item{with_conc_ratio}{Logical. If \code{TRUE}, includes the ratio of measured to target concentration in the results. Defaults to \code{TRUE}.}

\item{with_cv_intra}{Logical. If \code{TRUE}, includes intra-assay coefficient of variation (CV) for the in the results. Defaults to \code{TRUE}.}
}
\value{
A data frame containing the calibration results, including metrics such as bias, percentage bias, and intra-assay CV based on specified parameters.
}
\description{
This function retrieves calibration curve regression results from a \code{MidarExperiment} object.
It returns a summary of quality control (QC) metrics for specified QC samples.
including bias, absolute bias, and intra-assay coefficient of variation (CV).
The standard deviation of bias and percentage bias are also included unless the
it is \code{NA} for all analytes, i.e. when no replicates were measured.
}
\details{
The standard deviation of concentration is also included unless the number of replicates was 1.

The function uses data from the \code{MidarExperiment} object and filters it according to the specified QC types and other parameters. It then calculates summary statistics for each feature, such as bias and CV, and organizes the data into a user-specified format.
}
