% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-filtering.R
\name{get_response_curve_stats}
\alias{get_response_curve_stats}
\title{Linear Regression Statistics of Response Curves}
\usage{
get_response_curve_stats(
  data = NULL,
  with_staturation_stats = FALSE,
  limit_to_rqc = FALSE,
  silent_invalid_data = FALSE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object containing the dataset and response curve annotations.}

\item{with_staturation_stats}{Logical, if TRUE, include additional statistics from the \code{lancer} package.
Note: The \code{lancer} package must be installed when this argument is set to TRUE.}

\item{limit_to_rqc}{Logical, if TRUE (default), only include rows with \code{qc_type == "RQC"}.}

\item{silent_invalid_data}{Logical, if TRUE suppresses raising an error when
required data or metadata are missing, or there is a mismatch between them.}
}
\value{
A tibble with linear regression statistics (\code{r2}, \code{slopenorm}, \code{y0norm}) for each curve,
or \code{NULL} if no data matches the criteria.
}
\description{
This function calculates linear regression statistics (R-squared, slope, and intercept)
for each response curve in the provided \code{MidarExperiment} object. Optionally, it can include
additional statistics from the \code{lancer} package (if installed) when \code{with_staturation_stats} is set to TRUE.
}
