% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-filtering.R
\name{get_response_curve_stats}
\alias{get_response_curve_stats}
\title{Linear regression statistics of response curves}
\usage{
get_response_curve_stats(
  data = NULL,
  with_staturation_stats = FALSE,
  limit_to_rqc = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{with_staturation_stats}{Include statistics and classification from the \code{lancer} package.}

\item{limit_to_rqc}{If TRUE (default) then only includes RQC qc type}
}
\value{
Tibble with linear regression stats of all curves in a wide format
}
\description{
Linear regression statistics of response curves
}
\details{
When using \code{with_staturation_stats} then the \code{lancer} package needs to be installed
}
