% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{import_data_csv}
\alias{import_data_csv}
\title{Import plain analysis results}
\usage{
import_data_csv(
  data = NULL,
  path,
  variable_name,
  analysis_id_col = NA,
  import_metadata
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{One or more file names with path, or a folder path, which case all *.csv files in this folder will be read.}

\item{variable_name}{Variable type representing the values in the table. Must be one of "intensity", "norm_intensity", "conc", "area", "height", "response")}

\item{analysis_id_col}{Column to be used as analysis_id. \code{NA} (default) used 'analysis_id' if present, or the first column if it contains unique values.}

\item{import_metadata}{Import additional metadata columns (e.g. batch ID, sample type) and add to the \code{MidarExperiment} object.
Only following metadata column names are supported: "qc_type", "batch_id", "is_quantifier", "is_istd", "run_seq_num", "precursor_mz", "product_mz", "collision_energy"}
}
\value{
MidarExperiment object
}
\description{
Imports .csv file(s) reprensenting analysis results in a wide format with samples in rows and features in columns.
The first column must contain the analysis identifier, subsequent columns can be metadata columns, followed by value of different features.
The table format must be in the long format with columns for the raw data file name, feature ID, and the peak intensity and other arguments
Additional information, such as retention time, FWHM, precursor/product m/z, and CE will also be imported and will available from the \code{MidarExperiment} object for downstream analyses.

When a path to a folder is provided, all .scv files in that folder will be imported and merged into one raw dataset. This is useful, e.g. when importing datasets that were pre-processed in blocks resulting in different files.
Each feature and raw data file pair must only occur once within and across all .csv source data files, duplicated return an error.
}
\examples{
file_path <- system.file("extdata", "plain_wide_dataset.csv", package = "midar")

mexp <- MidarExperiment()

mexp <- import_data_csv(
  data = mexp,
  path = file_path,
 variable_name = "conc",
 import_metadata = TRUE)

print(mexp)

}
