% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{import_data_mrmkit}
\alias{import_data_mrmkit}
\title{Imports MRMkit peak integration results}
\usage{
import_data_mrmkit(data = NULL, path, import_metadata, silent = FALSE)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{One or more file paths, or a directory path (in which case all matching files will be imported)}

\item{import_metadata}{Logical, whether to import additional metadata columns (e.g., \code{batch_id}, \code{qc_type})}

\item{silent}{Logical, whether to suppress most notifications}
}
\value{
MidarExperiment object with the imported data
}
\description{
Imports tabular data files (*.tsv) generated from \code{MRMkit} containing peak integration results.
The input files must be in a long format with columns for the raw data file name, feature ID, peak intensity, and other parameters.
Additional information, such as retention time, FWHM, precursor/product MZ, and CE will also be imported and made available in the \code{MidarExperiment} object for downstream analyses.

When a directory path is provided, all matching files in that directory will be imported and merged into a single dataset.
This is useful when importing datasets that were pre-processed in blocks, resulting in multiple files.
Each unique combination of feature and raw data file must only occur once across all source data files.
Duplicate combinations will result in an error.
}
\examples{
mexp <- MidarExperiment()

file_path = system.file("extdata", "MRMkit_demo.tsv", package = "midar")

mexp <- import_data_mrmkit(
  data = mexp,
  path = file_path,
  import_metadata = TRUE)
print(mexp)
}
