% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{import_data_mrmkit}
\alias{import_data_mrmkit}
\title{Reads MRMkit results}
\usage{
import_data_mrmkit(data, file_dir_names)
}
\arguments{
\item{data}{MidarExperiment object}

\item{file_dir_names}{file path of MRMkit csv output with raw peak areas}

\item{silent}{No comments printed}
}
\value{
A tibble in the long format
}
\description{
Reads MRMkit results
}
