% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{import_metadata_msorganizer}
\alias{import_metadata_msorganizer}
\title{Import metadata from the MSOrganizer template (.XLM) and associates it with the analysis data}
\usage{
import_metadata_msorganizer(
  data,
  filename,
  analysis_sequence = c("timestamp", "resultfile", "metadata"),
  excl_unannotated_analyses = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{file name and path}

\item{excl_unannotated_analyses}{Exclude analyses (samples) that have no matching metadata}
}
\value{
MidarExperiment object
}
\description{
Requires version 1.9.1 of the template
}
