% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-metadata.R
\name{metadata_import_exceltemplate}
\alias{metadata_import_exceltemplate}
\title{Import metadata from the MSOrganizer template (.XLM) and associates it with the analysis data}
\usage{
metadata_import_exceltemplate(
  data,
  path,
  analysis_sequence = "default",
  excl_unannotated_analyses = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{file name and path}

\item{analysis_sequence}{Must by any of: "timestamp", "resultfile" or "metadata". Defines how the analysis order is determined. Default is "timestamp", when not available the sequence in the analysis results are used.}

\item{excl_unannotated_analyses}{Exclude analyses (samples) that have no matching metadata}
}
\value{
MidarExperiment object
}
\description{
Requires version 1.9.1 of the template
}
