% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-import.R
\name{import_metadata_msorganizer}
\alias{import_metadata_msorganizer}
\title{Import Metadata from a MIDAR Metadata Organizer file}
\usage{
import_metadata_msorganizer(
  data = NULL,
  path,
  ignore_warnings = FALSE,
  excl_unmatched_analyses = FALSE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object}

\item{path}{File name and path of the 'MIDAR Metadata Organizer' file (.xlsm/.xlsx) file}

\item{ignore_warnings}{Ignore warnings from data validation and proceed with importing metadata}

\item{excl_unmatched_analyses}{Exclude analyses (samples) that have no matching metadata}
}
\value{
An updated \code{MidarExperiment} object
}
\description{
Imports metadata from a 'MIDAR Metadata Organizer' file (.xlsm/.xlsx) file and associates it with analysis data.
}
\examples{
mexp <- MidarExperiment()

mexp <- import_data_mrmkit(
  data = mexp,
  path = system.file("extdata", "MRMkit_demo.tsv", package = "midar"),
  import_metadata = TRUE)

mexp <- import_metadata_msorganizer(
 data = mexp,
 path = system.file("extdata", "Example_Metadata_1.xlsm", package = "midar"),
 excl_unmatched_analyses = FALSE,
 ignore_warnings = TRUE)

print(mexp)

}
