% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-read.R
\name{metadata_from_data}
\alias{metadata_from_data}
\title{Retrieve available metadata from imported analysis data}
\usage{
metadata_from_data(data, qc_type_field = "qc_type")
}
\arguments{
\item{data}{MidarExperiment object}

\item{qc_type_field}{Column name in imported raw data file representing the \code{qc_type}}
}
\value{
MidarExperiment object
}
\description{
Available information will depend on the format of imported raw analysis data. See \code{\link[=data_import_agilent]{data_import_agilent()}} and \code{\link[=data_import_mrmkit]{data_import_mrmkit()}}
}
