% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-metadata.R
\name{metadata_from_data}
\alias{metadata_from_data}
\title{Retrieve available metadata from imported analysis data}
\usage{
metadata_from_data(data, analysis_sequence = "resultfile")
}
\arguments{
\item{data}{MidarExperiment object}

\item{analysis_sequence}{Must by any of: "resultfile" or "timestamp". Defines how the analysis order is according the sequence (default) in the loaded analysis data or the timestamp, if available, in the analysis data.}
}
\value{
MidarExperiment object
}
\description{
Available information will depend on the format of imported raw analysis data. See \code{\link[=rawdata_import_agilent]{rawdata_import_agilent()}} and \code{\link[=rawdata_import_mrmkit]{rawdata_import_mrmkit()}}
}
