% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-read.R
\name{import_metadata_midarxlm}
\alias{import_metadata_midarxlm}
\title{Import Metadata from MIDAR Template}
\usage{
import_metadata_midarxlm(
  data = NULL,
  path,
  excl_unmatched_analyses = FALSE,
  ignore_warnings = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object containing the analysis data}

\item{path}{File name and path of the MIDAR template (.XML) file}

\item{excl_unmatched_analyses}{Exclude analyses (samples) that have no matching metadata}

\item{ignore_warnings}{Ignore warnings from data validation and proceed with importing metadata}
}
\value{
The input MidarExperiment object with the metadata imported and associated
}
\description{
Imports metadata from a MIDAR template (.XML) file and associates it with  analysis data.
}
