% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-read.R
\name{metadata_import_midarxlm}
\alias{metadata_import_midarxlm}
\title{Import metadata from the midar template (.XLM) and associates it with the analysis data}
\usage{
metadata_import_midarxlm(data, path, ignore_warnings = FALSE)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{file name and path}

\item{excl_unannotated_analyses}{Exclude analyses (samples) that have no matching metadata}
}
\value{
MidarExperiment object
}
\description{
Requires version 1.9.1 of the template. Template is based on the MSorganizer template (see TODO)
}
