% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MidarExperiment-class}
\alias{MidarExperiment-class}
\title{S4 Class Representing the MIDAR Dataset}
\description{
Core class to store raw and processed data, annotations and metadata.
}
\details{
Provides also functions to validate data integrity, to perfom normalization and quantitation based on internal standards and to calculate the QC parameters of the experiment
}
\section{Slots}{

\describe{
\item{\code{analysis_type}}{Analysis type, one of "lipidomics", "metabolomics", "externalcalib", "others"}

\item{\code{dataset_orig}}{Original imported analysis data. Required fields:}

\item{\code{dataset}}{Processed analysis data. Required fields:}

\item{\code{dataset_QC_filtered}}{Processed analysis data. Required fields:}

\item{\code{annot_analyses}}{Annotation of analyses/runs}

\item{\code{annot_features}}{Annotation of measured features.}

\item{\code{annot_istd}}{Annotation of Internal Standard concs.}

\item{\code{annot_responsecurves}}{Annotation of  Response curves (RQC). Required fields}

\item{\code{annot_studysamples}}{Annotation of study samples. Required fields:}

\item{\code{annot_batch_info}}{Annotation of batches. Required fields:}

\item{\code{metrics_qc}}{QC information for each measured feature}

\item{\code{status_processing}}{Status within the data processing workflow}

\item{\code{is_istd_normalized}}{Flag if data has been ISTD normalized}

\item{\code{is_quantitated}}{Flag if data has been quantitated using ISTD and sample amount}

\item{\code{is_drift_corrected}}{Flag if data has been drift corrected}

\item{\code{is_batch_corrected}}{Flag if data has been batch corrected}

\item{\code{is_isotope_corr}}{Flag if one or more features have been isotope corrected}
}}

