% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-class.R
\docType{class}
\name{MidarExperiment-class}
\alias{MidarExperiment-class}
\title{S4 Class Representing the MIDAR Dataset}
\description{
Core class to store raw and processed data, annotations and metadata.
}
\details{
Provides also functions to validate data integrity, to perfom normalization and quantitation based on internal standards and to calculate the QC parameters of the experiment
}
\section{Slots}{

\describe{
\item{\code{dataset_orig}}{Original imported analysis data. Required fields:}

\item{\code{dataset}}{Processed analysis data. Required fields:}

\item{\code{dataset_QC_filtered}}{Processed analysis data. Required fields:}

\item{\code{annot_analyses}}{Annotation of analyses/runs}

\item{\code{annot_features}}{Annotation of measured features.}

\item{\code{annot_istd}}{Annotation of Internal Standard concentrations.}

\item{\code{annot_responsecurves}}{Annotation of  Response curves (RQC). Required fields}

\item{\code{annot_studysamples}}{Annotation of study samples. Required fields:}

\item{\code{annot_batch_info}}{Annotation of batches. Required fields:}

\item{\code{d_QC}}{QC information for each measured feature}

\item{\code{status_processing}}{Status within the data processing workflow}
}}

