% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-class.R
\name{MidarExperiment}
\alias{MidarExperiment}
\title{Constructor for the MidarExperiment object.}
\usage{
MidarExperiment()
}
\value{
\code{MidarExperiment} object
}
\description{
Constructor for the MidarExperiment object.
}
