% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-processing.R
\name{normalize_by_istd}
\alias{normalize_by_istd}
\title{Normalize Intensities with corresponding ISTD Intensities}
\usage{
normalize_by_istd(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
MidarExperiment object
}
\description{
Normalize Intensities with corresponding ISTD Intensities
}
