% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istd-normalization.R
\name{normalize_by_istd}
\alias{normalize_by_istd}
\title{Normalize Intensities with corresponding ISTD Intensities}
\usage{
normalize_by_istd(data, interference_correction = FALSE)
}
\arguments{
\item{data}{MidarExperiment object}

\item{interference_correction}{Apply interference (e.g. isotope) correction as defined in the feature metadata}
}
\value{
MidarExperiment object
}
\description{
Normalize Intensities with corresponding ISTD Intensities
}
