% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istd-normalization.R
\name{normalize_by_istd}
\alias{normalize_by_istd}
\title{Normalize Intensities with corresponding ISTD Intensities}
\usage{
normalize_by_istd(data, error_missing_info = TRUE)
}
\arguments{
\item{data}{MidarExperiment object}

\item{error_missing_info}{Raise error when ISTD is not defined for 1 or more features (excluding ISTD themselves) . Default: \code{TRUE}.}

\item{interference_correction}{Apply interference (e.g. isotope) correction as defined in the feature metadata}
}
\value{
MidarExperiment object
}
\description{
Normalize Intensities with corresponding ISTD Intensities
}
