% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata-import.R
\name{parse_mrmkit_result}
\alias{parse_mrmkit_result}
\title{Parses MRMkit peak integration results into a tibble}
\usage{
parse_mrmkit_result(path, use_normalized_data = FALSE, silent = FALSE)
}
\arguments{
\item{path}{File name of the MRMkit result file (*.tsv or *.csv)}

\item{use_normalized_data}{Import raw peak areas or normalized peak areas from the file}

\item{silent}{No comments printed}
}
\value{
A tibble in the long format
}
\description{
Parses MRMkit peak integration results into a tibble
}
