% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-pca.R
\name{plot_pca_qc}
\alias{plot_pca_qc}
\title{PCA plot for QC}
\usage{
plot_pca_qc(
  data,
  variable,
  dim_x,
  dim_y,
  log_transform,
  remove_istds,
  point_size = 5,
  point_alpha = 0.8,
  ellipse_alpha = 0.8,
  font_base_size = 12
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable}{which variable to use for plot}

\item{dim_x}{PCA dimension on x axis}

\item{dim_y}{PCA dimension on y axis}

\item{log_transform}{log transform data for plot}

\item{remove_istds}{exclude internal standards}

\item{point_size}{size of points}

\item{point_alpha}{transparency of points}

\item{ellipse_alpha}{transparency of ellipse fill}

\item{font_base_size}{Base font size for plot text elements}
}
\value{
ggplot2 object
}
\description{
PCA plot for QC
}
