% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_QC.R
\name{plot_qc_summary_classes}
\alias{plot_qc_summary_classes}
\title{Plot results of QC filter per analyte class}
\usage{
plot_qc_summary_classes(data, user_defined_keeper = FALSE, base_size = 8)
}
\arguments{
\item{data}{MidarExperiment object}

\item{user_defined_keeper}{Include user-defined feature inclusion list, even they did not pass QC filtering}

\item{base_size}{font size of plots}
}
\value{
ggplot2 object
}
\description{
Plot results of QC filter per analyte class
}
