% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_qc.R
\name{plot_responsecurves}
\alias{plot_responsecurves}
\title{Response curves plot}
\usage{
plot_responsecurves(
  data,
  use_filt_data,
  output_pdf,
  response_variable = "feature_intensity",
  feature_incl_filt = "",
  feature_excl_filt = "",
  regr_max_percent = NA,
  pdf_filename = "",
  rows_page = 4,
  columns_page = 5,
  point_size = 2,
  line_width = 1,
  text_scale_factor = 1,
  return_plot_list = FALSE,
  base_size = 7
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{use_filt_data}{Use QC-filtered data}

\item{output_pdf}{Save as PDF}

\item{response_variable}{Variable to plot}

\item{feature_incl_filt}{Filter features names matching the criteria (regex). When empty, \code{NA} or \code{NULL} all available features are included.}

\item{feature_excl_filt}{Exclude features names matching the criteria (regex).  When empty, \code{NA} or \code{NULL} all available features are included.}

\item{regr_max_percent}{Max relative sample amount to use in regressionb}

\item{pdf_filename}{file name of pdf file}

\item{rows_page}{rows per page}

\item{columns_page}{columns per page}

\item{point_size}{point size}

\item{line_width}{regression line width}

\item{text_scale_factor}{text scale factor}

\item{return_plot_list}{return plot as list}

\item{base_size}{base font size}
}
\value{
A list of ggplot2 objects
}
\description{
Response curves plot
}
