% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-plots_QC.R
\name{plot_responsecurves}
\alias{plot_responsecurves}
\title{Response curves plot}
\usage{
plot_responsecurves(
  data,
  output_PDF,
  response_variable = "Intensity",
  include_features_containing = "",
  exclude_features_containing = "",
  regr_max_percent = NA,
  pdf_file_name = "",
  rows_page = 4,
  columns_page = 5,
  point_size = 2,
  line_width = 1,
  text_scale_factor = 1,
  return_plot_list = FALSE,
  base_size = 7
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{output_PDF}{Save as PDF}

\item{response_variable}{Variable to plot}

\item{include_features_containing}{Filter features containing}

\item{exclude_features_containing}{Exclude features containing}

\item{regr_max_percent}{Max relative sample amount to use in regressionb}

\item{pdf_file_name}{file name of pdf file}

\item{rows_page}{rows per page}

\item{columns_page}{columns per page}

\item{point_size}{point size}

\item{line_width}{regression line width}

\item{text_scale_factor}{text scale factor}

\item{return_plot_list}{return plot as list}

\item{base_size}{base font size}
}
\value{
A list of ggplot2 objects
}
\description{
Response curves plot
}
