% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-runscatter.R
\name{plot_runscatter}
\alias{plot_runscatter}
\title{RunScatter Plot}
\usage{
plot_runscatter(
  data = NULL,
  variable = c("intensity", "norm_intensity", "conc", "rt", "area", "height", "fwhm",
    "width", "symmetry", "intensity_raw", "intensity_before", "norm_intensity_raw",
    "norm_intensity_before", "conc_raw", "conc_before"),
  filter_data = FALSE,
  qc_types = NA,
  include_qualifier = TRUE,
  include_istd = TRUE,
  include_feature_filter = NA,
  exclude_feature_filter = NA,
  plot_range = NA,
  output_pdf = FALSE,
  path = NA,
  multithreading = FALSE,
  return_plots = FALSE,
  show_batches = TRUE,
  batch_zebra_stripe = FALSE,
  batch_line_color = "#cdf7d9",
  batch_fill_color = "grey93",
  cap_outliers = FALSE,
  cap_sample_k_mad = 4,
  cap_qc_k_mad = 4,
  cap_top_n_outliers = NA,
  show_reference_lines = FALSE,
  ref_qc_types = NA,
  reference_k_sd = 2,
  reference_batchwise = FALSE,
  reference_line_color = "#04bf9a",
  reference_sd_shade = FALSE,
  reference_fill_color = NA,
  reference_linewidth = 0.75,
  show_trend = FALSE,
  trend_color = "#22e06b",
  y_lim = c(0, NA),
  log_scale = FALSE,
  show_gridlines = FALSE,
  point_size = 1.5,
  point_transparency = 1,
  point_border_width = NA,
  base_font_size = 10,
  rows_page = 3,
  cols_page = 3,
  specific_page = NA,
  page_orientation = "LANDSCAPE",
  y_label_text = NA,
  pages_per_core = 1,
  use_dingbats = TRUE,
  show_progress = TRUE
)
}
\arguments{
\item{data}{A \code{MidarExperiment} object containing the dataset and metadata.}

\item{variable}{The variable to plot on the y-axis, one of 'intensity',
'norm_intensity', 'conc', 'conc', 'rt', 'fwhm', 'area', 'height', response'.
Add \verb{_before} after the variable name to plot the feature values before the last
applied drift/batch correction, (e.g., \code{conc_before}). Add \verb{_raw} after the
variable name to plot the raw uncorrected feature values (e.g., \code{conc_raw}).}

\item{filter_data}{Logical, whether to use QC-filtered data based on criteria set via \code{filter_features_qc()}.}

\item{qc_types}{QC types to be plotted. Can be a vector of QC types or a regular expression pattern. \code{NA} (default) displays all available QC/Sample types.}

\item{include_qualifier}{Logical, whether to include qualifier features. Default is \code{TRUE}.}

\item{include_istd}{Logical, whether to include internal standard (ISTD) features. Default is \code{TRUE}.}

\item{include_feature_filter}{A regex pattern or a vector of feature names used to filter features by \code{feature_id}.
If \code{NA} or an empty string (\code{""}) is provided, the filter is ignored. When a vector of length > 1 is supplied,
is supplied, only features with exactly these names are selected (applied individually as OR conditions).}

\item{exclude_feature_filter}{A regex pattern or a vector of feature names to exclude features by feature_id.
If \code{NA} or an empty string (\code{""}) is provided, the filter is ignored. When a vector of length > 1 is supplied,
is supplied, only features with exactly these names are excluded (applied individually as OR conditions).}

\item{plot_range}{Numeric vector of length 2, specifying the start and end indices of the analysis order to be plotted. \code{NA} plots all samples.}

\item{output_pdf}{Logical, whether to save the plot as a PDF file.}

\item{path}{File name for the PDF output.}

\item{multithreading}{Logical, whether to use parallel processing to speed up plot generation.}

\item{return_plots}{Logical, whether to return the list of ggplot objects.}

\item{show_batches}{Logical, whether to show batch separators in the plot.}

\item{batch_zebra_stripe}{Logical, whether to display batches with alternating shaded and non-shaded areas.}

\item{batch_line_color}{Color of the batch separator lines.}

\item{batch_fill_color}{Color for the shaded areas representing batches.}

\item{cap_outliers}{Logical, whether to cap upper outliers based on MAD fences of SPL and QC samples.}

\item{cap_sample_k_mad}{Numeric, k * MAD (median absolute deviation) for outlier capping of SPL samples.}

\item{cap_qc_k_mad}{Numeric, k * MAD (median absolute deviation) for outlier capping of QC samples.}

\item{cap_top_n_outliers}{Numeric, cap the top n outliers regardless of MAD fences. \code{NA} or \code{0} ignores this filter.}

\item{show_reference_lines}{Whether to display reference lines (mean \eqn{\pm} n x SD).}

\item{ref_qc_types}{QC type for which the reference lines are calculated.}

\item{reference_k_sd}{Multiplier for standard deviations to define SD reference lines.}

\item{reference_batchwise}{Whether to calculate reference lines per batch.}

\item{reference_line_color}{Color of the reference lines.}

\item{reference_sd_shade}{\code{TRUE} plots a colored band indicating the \eqn{\pm} n x SD
reference range. \code{FALSE} (default) shows reference lines instead.}

\item{reference_fill_color}{Fill color of the batch-wise reference ranges.
If \code{NA} (default), the color assigned to the qc_type is used.}

\item{reference_linewidth}{Width of the reference lines.}

\item{show_trend}{If \code{TRUE} trend curves before or after drift/batch correction are shown.}

\item{trend_color}{Color of the trend curve.}

\item{y_lim}{Numeric vector of length 2, specifying the lower and upper y-axis limits. Default is `c(0,NA)``, which sets the lower limit to 0 and the upper limit automatically.}

\item{log_scale}{Logical, whether to use a log10 scale for the y-axis.}

\item{show_gridlines}{Whether to show major x and y gridlines.}

\item{point_size}{Size of the data points. Default is \code{1.5}.}

\item{point_transparency}{Alpha transparency of the data points.}

\item{point_border_width}{Width of the data point borders.}

\item{base_font_size}{Base font size for the plot.}

\item{rows_page}{Number of rows per page.}

\item{cols_page}{Number of columns per page.}

\item{specific_page}{Show/save a specific page number only. \code{NA} plots/saves all pages.}

\item{page_orientation}{Page orientation, "LANDSCAPE" or "PORTRAIT".}

\item{y_label_text}{Override the default y-axis label text.}

\item{pages_per_core}{Number of pages to be plotted by core when multithreading is enabled. Default is \code{6}. Changing this number may improve performance.}

\item{use_dingbats}{Logical, whether to use Dingbats font in the PDF output for improved plotting speed. Default is \code{TRUE}. Set to \code{FALSE} if your PDF viewer does not show points correctly.}

\item{show_progress}{Logical, whether to show a progress bar. Default is \code{TRUE}.}
}
\value{
A list of ggplot2 plots, or \code{NULL} if `return
}
\description{
The \code{runscatter} function visualizes raw or processed feature signals across
different sample/QC types along the analysis sequence. It helps identify
trends, detect outliers, and assess analytical performance. Available
feature variables, such as retention time (RT) and full width at
half maximum (FWHM), can be plotted against analysis order or timestamps.

By default, all QC types present in the dataset will be plotted. QC types that
predefined colors or shapes are assigned black shapes.
User-defined QC types that have no predefined colors or shapes in midar.
will be assigned black shapes.
have no predefined color and shape, will be assigned shapes in black. To show specific
QC types use the \code{qc_types} argument.

To plot the feature values before the last applied drift/batch correction,
add \verb{*_before} to the variable name, e.g., \code{intensity_before} or \code{conc_before}.
To plot the uncorrected feature values (before any drift/batch correction),
add \verb{*_raw} to the variable name, e.g., \code{intensity_raw} or \code{conc_raw}. To show
corresponding fit curves, set \code{show_trend = TRUE}.

The function also supports visualizing analysis batches, reference lines
(mean \eqn{\pm} SD), and trends. It offers customization options to display batch
separators, apply outlier capping, show smoothed trend curves, add reference
lines, and incorporate other features. Outlier capping is particularly useful
to focus on QC or study sample trends that might otherwise be obscured by
extreme values or high variability.

The \code{runscatter} function serves as a central QC tool in the workflow,
providing critical insights into data quality.
}
\details{
\itemize{
\item The outlier capping feature (\code{cap_outliers}) allows you to cap upper outliers
based on median absolute deviation (MAD) fences of SPL and QC samples, or to
remove the top n points. This can help to focus on the trends of interest when
there are outlier or a high variability in the data, e.g. in the study samples.
\item When using log-scale (\code{log_scale = TRUE}), zero or negative values will
replaced with the minimum positive value divided by 5 to avoid log 0 errors
\item Reference lines/ranges corresponding to mean \eqn{\pm} k x SD can be shown across
or within batches as lines or shaded stripes.
\item Trend curves can be displayed before or after drift/batch correction. In
either case, a drift and/or batch correction must be applied to the data
to enable plotting of trend curves. To show trend curves used for the last drift or batch correction,
add "_before" to the variable name, e.g. \code{conc_before} or \code{intensity_before}
and set \code{show_trend = TRUE}.
}
}
