% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-filtering.R
\name{qc_calc_metrics}
\alias{qc_calc_metrics}
\title{Calculate feature quality control (QC) metrics}
\usage{
qc_calc_metrics(
  data = NULL,
  batch_medians,
  with_norm_intensity = TRUE,
  with_conc = TRUE,
  with_linearity = TRUE,
  with_calibration = TRUE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{batch_medians}{Use median of batch-wise derived QC values. Default is FALSE.}

\item{with_norm_intensity}{Include normalized intensity statistics of features. Default is TRUE.}

\item{with_conc}{Include concentration statistics of features. Default is TRUE.}

\item{with_linearity}{Include linearity statistics of response curves. This will increase the calculation time. Default is TRUE.}

\item{with_calibration}{Include external calibration results if available.}
}
\value{
MidarExperiment object
}
\description{
For each feature, different QC metrics are calculated, either across the
full analysis or as medians of batch-wise calculations
}
