% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-filtering.R
\name{qc_calculate_metrics}
\alias{qc_calculate_metrics}
\title{Calculate feature quality control (QC) metrics}
\usage{
qc_calculate_metrics(data, batchwise_median)
}
\arguments{
\item{data}{MidarExperiment object}

\item{batchwise_median}{Use median of batch-wise derived QC values.}
}
\value{
MidarExperiment object
}
\description{
For each feature, different QC metrics are calculated, either across the
full analysis or as medians of batch-wise calculations
}
