% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-responsecurves.R
\name{qc_plot_responsecurves}
\alias{qc_plot_responsecurves}
\title{Response curves plot}
\usage{
qc_plot_responsecurves(
  data,
  response_variable = "feature_intensity",
  qc_filter_data,
  feature_incl_filt = "",
  feature_excl_filt = "",
  save_pdf = FALSE,
  path = "",
  regr_max_percent = NA,
  rows_page = 4,
  columns_page = 5,
  page_no = NA,
  point_size = 2,
  line_width = 1,
  text_scale_factor = 1,
  paper_orientation = "LANDSCAPE",
  base_size = 7,
  show_progress = TRUE,
  return_plot_list = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{response_variable}{Variable to plot}

\item{qc_filter_data}{Use QC-filtered data}

\item{feature_incl_filt}{Filter features names matching the criteria (regex). When empty, \code{NA} or \code{NULL} all available features are included.}

\item{feature_excl_filt}{Exclude features names matching the criteria (regex).  When empty, \code{NA} or \code{NULL} all available features are included.}

\item{save_pdf}{Save as PDF}

\item{path}{file name of pdf file}

\item{regr_max_percent}{Max relative sample amount to use in regressionb}

\item{rows_page}{rows per page}

\item{columns_page}{columns per page}

\item{page_no}{Specific page to plot. Default \code{NA}, meaning all pages are plotted}

\item{point_size}{point size}

\item{line_width}{regression line width}

\item{text_scale_factor}{text scale factor}

\item{paper_orientation}{Landscape/Portrait}

\item{base_size}{base font size}

\item{show_progress}{show progress bar}

\item{return_plot_list}{return plot as list}
}
\value{
A list of ggplot2 objects
}
\description{
Response curves plot
}
