% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-runorder.R
\name{qc_plot_rla_boxplot}
\alias{qc_plot_rla_boxplot}
\title{RLA (Relative Log Abundance) plot}
\usage{
qc_plot_rla_boxplot(
  data,
  rla_type_batch = c("within", "across"),
  variable = c("intensity", "norm_intensity", "conc", "conc_raw", "area", "height",
    "fwhm"),
  qc_filter_data = FALSE,
  qc_types = NA,
  feature_incl_filt = "",
  feature_excl_filt = "",
  analysis_no_range = NA,
  min_feature_intensity = 0,
  y_lim = NA,
  ignore_outliers = FALSE,
  show_batches = TRUE,
  batches_as_shades = FALSE,
  batch_line_color = "#b6f0c5",
  batch_shading_color = "grey93",
  x_axis_variable = c("run_seq_num"),
  x_gridlines = FALSE,
  linewidth = 0.2,
  base_font_size = 8,
  relative_log_abundances = TRUE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{rla_type_batch}{Must be either \code{within} or \code{across}, defining whether to use with-in or across-group RLA}

\item{variable}{Variable to plot}

\item{qc_filter_data}{Use QC-filtered data}

\item{qc_types}{QC type to plot. When qc_types is NA or NULL, all available QC types are plotted.}

\item{feature_incl_filt}{Filter text to select specific features (regex string)}

\item{feature_excl_filt}{Filter text to exclude specific features (regex string)}

\item{analysis_no_range}{Analysis range to plot. Format: c(start, end). Setting one of them to NA ignoresthe corresponding boundary. Default is NA, which plots all available analyses.}

\item{min_feature_intensity}{Exclude features with overall median signal below this value}

\item{y_lim}{Y-axis lower and upper limits as vector (default is NA). This also overwrites limits calculated by \code{ignore_outliers}}

\item{ignore_outliers}{Exclude outlier values based on 4x MAD (median absolute deviation) fences}

\item{show_batches}{Show batches}

\item{batches_as_shades}{Show batches as shades}

\item{batch_line_color}{batch separator color}

\item{batch_shading_color}{batch shade color}

\item{x_gridlines}{Show major x gridlines}

\item{linewidth}{Line width used for whiskers of boxplot}

\item{base_font_size}{base font size for plots (default is 8)}

\item{relative_log_abundances}{Use relative log abundances (RLA). If \code{FALSE} just use log-transformed values (then the result is not an RLA plot)}

\item{feature_filter}{Filter text to select specific features (regex string)}
}
\value{
ggplot object
}
\description{
Relative log abundance (RLA) plots show standardized analyte abundances of each sample. Standardization is done by removing the within- or across-batch median from each analyte.
}
\details{
This plot was first introduced by De Livera et al. (2012) in the context of metabolomics and
is used to visualize the relative log abundance (RLA) of each feature in each sample.
The RLA is calculated by subtracting the within- or across-batch median from each feature.
This normalization is useful to examine the dataset for technical effect that involve all feature in a similar manner,
such as batch effects due to changes of instrument response, pipeting errors,
sample spillage.
As all features are normmalized, this plot is more robust to detect such effects,
unlike e.g. sum abundance plots which are a biased toward the most abundant lipid species in the analysed matrix
}
\references{
De Livera et al. (2012) Normalizing and integrating metabolomics data. Analytical Chemistry 10768-10776
\href{https://doi.org/10.1021/ac302748b}{DOI: 10.1021/ac302748b}

De Livera et al. (2015) Statistical Methods for Handling Unwanted Variation in Metabolomics Data. Analytical Chemistry 87(7):3606-3615
\href{https://doi.org/10.1021/ac502439y}{DOI: 10.1021/ac502439y}
}
