% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-qc-runorder.R
\name{qc_plot_runscatter}
\alias{qc_plot_runscatter}
\title{RunScatter plot}
\usage{
qc_plot_runscatter(
  data,
  variable = c("intensity", "norm_intensity", "conc", "conc_raw", "area", "height",
    "fwhm"),
  qc_filter_data = FALSE,
  qc_types = NA,
  filt_include_features = NA,
  filt_exclude_features = NA,
  analysis_no_range = NA,
  save_pdf = FALSE,
  path = "",
  log_scale = FALSE,
  cap_outliers = FALSE,
  cap_qc_k_mad = 4,
  cap_spl_k_mad = 4,
  cap_top_n_values = NA,
  show_batches = TRUE,
  show_control_limits = NA,
  set_limits_n_sd = NA,
  limits_batchwise = FALSE,
  limits_linecolor = "#38dff5",
  limits_linewidth = 0.75,
  batches_as_shades = FALSE,
  batch_line_color = "#b6f0c5",
  batch_shading_color = "grey93",
  cols_page = 3,
  rows_page = 3,
  base_font_size = 11,
  annot_scale = 1,
  show_trend = FALSE,
  trend_linecolor = "#22e06b",
  fit_qc_type = "BQC",
  paper_orientation = "LANDSCAPE",
  point_transparency = 1,
  point_size = 2,
  point_stroke_width = 1,
  page_no = NA,
  y_label_text = NA,
  return_plot_list = FALSE,
  show_gridlines = FALSE,
  show_progress = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{variable}{Variable to plot}

\item{qc_filter_data}{Use QC-filtered data, based on criteria set via \code{qc_set_feature_filters()}}

\item{qc_types}{QC type to plot. When qc_types us NA or NULL, all available QC types are plotted.}

\item{filt_include_features}{Select features with feature_id matching the given string. By default a \code{regex} string. \code{NA}, \code{""} ignores the filter.}

\item{filt_exclude_features}{Exclude features with feature_id matching the given string. By default a \code{regex} string. \code{NA}, \code{""} ignores the filter.}

\item{analysis_no_range}{Analysis range to plot. Format: c(start, end). Setting one of them to NA ignoresthe corresponding boundary. Default is NA, which plots all available analyses.}

\item{save_pdf}{save as PDF}

\item{path}{file name of PDF}

\item{log_scale}{Use log10 scale for the y axis}

\item{cap_outliers}{Cap upper outliers, based on MAD fences of SPL and QC samples, see \code{cap_spl_k_mad} and \code{cap_qc_k_mad}. Useful to cap extreme values distorting the plot.}

\item{cap_qc_k_mad}{k * MAD (median absolute deviation) for outlier capping of BQC and/or TQC samples. Default is 4.}

\item{cap_spl_k_mad}{k * MAD (median absolute deviation) for outlier capping of SPL samples. Default is 4.}

\item{cap_top_n_values}{Cap top n values, irrespective of the IQR fence. This is useful to remove single or few extreme values. Default (\code{NA}) or \code{0} ignores the filter.}

\item{show_batches}{Show batches}

\item{show_control_limits}{Show mean line of the given QC types. NA (default) ignores the median line.}

\item{set_limits_n_sd}{Show  /- n x SD lines of the QC types defined via \code{show_median}. \code{NA} (default) ignores the SD lines.}

\item{limits_batchwise}{Calculate mean and SD line per batch. Default is \code{FALSE}.}

\item{limits_linecolor}{Color of mean and SD lines}

\item{limits_linewidth}{Width of mean and SD lines}

\item{batches_as_shades}{Show batches as shades}

\item{batch_line_color}{batch separator color}

\item{batch_shading_color}{batch shade color}

\item{cols_page}{columns per page}

\item{rows_page}{rows per page}

\item{base_font_size}{base font size of plot}

\item{annot_scale}{scale factor of text elements}

\item{show_trend}{Show drift correction. TODO: Add more details}

\item{trend_linecolor}{Color of tend line}

\item{fit_qc_type}{QC type used for smoothing  TODO: Add more detail}

\item{paper_orientation}{Landscape/Portrait}

\item{point_transparency}{Alpha of points}

\item{point_size}{point size}

\item{point_stroke_width}{point stroke width}

\item{page_no}{Show/save specific page number. Default is \code{NA}, which plots all pages.}

\item{y_label_text}{Overwrite y label with this text}

\item{return_plot_list}{return list with plots}

\item{show_gridlines}{show x and y major gridlines}

\item{show_progress}{show progress bar}
}
\value{
A list of ggplot2 plots or NULL
}
\description{
RunScatter plot
}
