% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-istd-normalization.R
\name{quantify_by_istd}
\alias{quantify_by_istd}
\title{Calculate Analyte Concentrations Based on Internal Standards}
\usage{
quantify_by_istd(data = NULL, fail_missing_annotation = TRUE)
}
\arguments{
\item{data}{A \code{MidarExperiment} object}

\item{fail_missing_annotation}{If \code{TRUE}, an error will be raised if any of the following information is missing: ISTD concentration, ISTD mix volume, and sample amounts for any feature.
If \code{FALSE}, missing annotations will be ignored, and resulting feature concentration will be \code{NA}}
}
\value{
A \code{MidarExperiment} object with the calculated analyte concentrations added to the
\code{dataset} table in the \code{feature_conc} column.
}
\description{
Analyte concentrations are calculated based on ISTD-normalized intensities
and the corresponding spiked-in ISTD amount, normalized by the sample amount
(single-point calibration approach). The concentration unit is determined
from the \code{sample_amount_unit} field in the analysis metadata of the \code{MidarExperiment} object.
}
\details{
The calculated concentrations are added to the \code{dataset} table as a new column
named \code{feature_conc}.
}
\seealso{
\code{\link[=quantify_by_calibration]{quantify_by_calibration()}} for calculating concentrations based on external calibration curves.
}
