% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/istd-normalization.R
\name{quantitate_by_istd}
\alias{quantitate_by_istd}
\title{Quantitate using sample and spiked ISTD amounts}
\usage{
quantitate_by_istd(data)
}
\arguments{
\item{data}{MidarExperiment object}
}
\value{
MidarExperiment object
}
\description{
Quantitate using sample and spiked ISTD amounts
}
