% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawdata-import.R
\name{rawdata_import_agilent}
\alias{rawdata_import_agilent}
\title{Imports an Agilent MassHunter Quant CSV file}
\usage{
rawdata_import_agilent(
  data,
  path,
  use_metadata,
  file_format = "csv",
  expand_qualifier_names = TRUE,
  silent = FALSE
)
}
\arguments{
\item{data}{MidarExperiment object}

\item{path}{One or more file names with path, or a folder path, which case all *.csv files in this folder will be read.}

\item{use_metadata}{Extract and add metadata from the analysis result file}

\item{file_format}{File format of the MassHunter export. One of "csv", "xls"}

\item{expand_qualifier_names}{Add quantifier name in front of qualifier name, as the latter only has the mz transitions values}

\item{silent}{Suppress most notifications}
}
\value{
MidarExperiment object
}
\description{
Imports .csv file(s) exported from \verb{Agilent MassHunter (MH) Quantitative Analysis} containing peak integration results.
Samples must be in rows, features/compounds in columns and must contain either peak areas, peak heights or intensities.
Additional columns, such as RT (retention time), FWHM, PrecursorMZ, and CE will be imported and will available from the \code{MidarExperiment} object for downstream analyses.

When a path to a folder is provided, all .csv files in that folder will be imported and merged into one raw dataset. This is useful, e.g. when importing datasets that are pre-processing in blocks resulting in different files.
Each feature and raw data file pair must only occur once within and across all .csv source data files, duplicated return an error.
}
\examples{
file_path <- system.file("extdata", "Example_MHQuant_1.csv", package = "midar")
mexp <- MidarExperiment()
mexp <- rawdata_import_agilent(data = mexp, path = file_path, use_metadata = TRUE)
mexp
}
