% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_masshunter_csv}
\alias{read_masshunter_csv}
\title{Reads and parses one Agilent MassHunter Quant CSV result file}
\usage{
read_masshunter_csv(filename, expand_qualifier_names = FALSE, silent = FALSE)
}
\arguments{
\item{filename}{File path of MassHunter Quant CSV file}

\item{expand_qualifier_names}{If TRUE, original qualifier names will be renamed by adding the quantifier name in front and placing qualifier name into square brackets(e.g. \code{Qualifier (422.3 -> 113.0)} ransition names of quantifier will be added to qualifier names}

\item{silent}{Suppress messages}
}
\value{
A tibble with the parse results in the long format
}
\description{
Reads and parses one Agilent MassHunter Quant CSV result file
}
