% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdata.R
\name{read_mrmkit_result}
\alias{read_mrmkit_result}
\title{Reads a long CSV file with Feature Intensities}
\usage{
read_mrmkit_result(filename, use_normalized_data = FALSE, silent = FALSE)
}
\arguments{
\item{silent}{No comments printed}

\item{data}{MidarExperiment object}

\item{file_dir_names}{One or more file names with path, or a folder path, which case all *.csv or *.tsv files in this folder will be read.}

\item{final_results_file}{file path of MRMkit csv output with final processed normalized peak areas}
}
\value{
A tibble in the long format
}
\description{
Reads a long CSV file with Feature Intensities
}
