% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miqar-class.R
\name{read_msorganizer_xlm}
\alias{read_msorganizer_xlm}
\title{Import metadata from the MSOrganizer template (.XLM)}
\usage{
read_msorganizer_xlm(data, filename)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{file name and path}
}
\value{
MidarExperiment object
}
\description{
Import metadata from the MSOrganizer template (.XLM)
}
