% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{read_msorganizer_xlm}
\alias{read_msorganizer_xlm}
\title{Import metadata from the MSOrganizer template (.XLM)}
\usage{
read_msorganizer_xlm(data, filename, excl_unannotated_analyses = FALSE)
}
\arguments{
\item{data}{MidarExperiment object}

\item{filename}{file name and path}

\item{excl_unannotated_analyses}{Exclude analyses (samples) that have no matching metadata}
}
\value{
MidarExperiment object
}
\description{
Import metadata from the MSOrganizer template (.XLM)
}
