% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-managment.R
\name{set_analysis_order}
\alias{set_analysis_order}
\title{Set the analysis order}
\usage{
set_analysis_order(data = NULL, order_by = "default")
}
\arguments{
\item{data}{MidarExperiment object}

\item{order_by}{Must by any of: "timestamp", "resultfile" or "metadata". Defines how the analysis order is determined. Default is "timestamp", when not available the sequence in the analysis results are used.}
}
\value{
MidarExperiment object
}
\description{
Sets the analysis order (sequence), based on either (i) analysis timestamp, if available, (ii) the order in which analysis appeared in the imported raw data file, or (iii) the order in which analyses were defined in the Analysis metadata
}
\examples{
file_path <- system.file("extdata", "sPerfect_MRMkit.tsv", package = "midar")
mexp <- MidarExperiment()
mexp <- import_data_mrmkit(mexp, path = file_path, include_metadata = TRUE)
mexp <- set_analysis_order(mexp, "timestamp")
}
