% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-managment.R
\name{set_intensity_var}
\alias{set_intensity_var}
\title{Set default variable to be used as feature raw signal value}
\usage{
set_intensity_var(data, variable_name, auto_select = FALSE, ...)
}
\arguments{
\item{data}{MidarExperiment object}

\item{analysis_sequence}{Must by any of: "timestamp", "resultfile" or "metadata". Defines how the analysis order is determined. Default is "timestamp", when not available the sequence in the analysis results are used.}
}
\value{
MidarExperiment object
}
\description{
Sets the raw signal variable used for calculations starting from raw signal
values (i.e., normalization)
}
